R documentation for ClinReport package

It enables to create easily a statistical reporting in Microsoft Word documents with tables in a pretty format according to "clinical standards".

Details

It can also be used outside the scope of clinical trials, for any statistical reporting in Word.

Descriptive tables for quantitative statistics (mean, median, max etc..) and qualitative statistics (frequencies and percentages) are available and pretty tables of Least Square Means of LM, LME and GLM models coming from emmeans function in emmeans package are also available.

The package works with officer and flextable packages to export the outputs into Microsoft Word documents.

There are 2 main functions for descriptive reporting : report.quanti and report.quali

There is 1 unique function for model LS means reporting: report.lsmeans

Each of those 3 functions creates a desc object that can be used with report.doc to export the result(s) into a Word document. This is the unique function to export the output(s) into a Word document.

Future versions will improve the "standard" graphic reporting, the anova table reporting and statistics tables specific for Adverse Events reporting. Survival models will also be included.