Constructor function for the desc object

Creates a desc object.

desc(output = NULL, total = NULL, nbcol = NULL, y = NULL,
  x1 = NULL, x2 = NULL, at.row = NULL, subjid = NULL,
  type.desc = NULL, type = NULL, y.label = NULL, type.mod = NULL,
  raw.output = NULL, contrast = NULL, contrast.name = NULL,
  stat.name = NULL)

Arguments

output
a data.frame object

total
Boolean. Indicates if a column Total should be added or not

nbcol
Numeric. number of columns that are not statistics

y
Character indicating the response variable

x1
Character indicating a factor in the data

x2
Character indicating a factor in the data

at.row
Character. Pass to spacetable function. Used to space the results (see spacetable)

subjid
Character. Indicates in the data.frame the name of the column used to identify the Id of the subjects. If not null, it adds in the headers the number of unique subject per levels of x1 or y (if x1 is null).

type.desc
Character. Can be "quali" "quanti" "lsmeans" or "quali_quanti"

type
Character. Can be "response" or "link"

y.label
Character. The label for y response variable

type.mod
Character. The type of models the LS means are coming from (can be 'quali' or 'quanti')

raw.output
Data.frame. The raw statistics unformatted from report.quali, report.quanti and report.lsmeans

contrast
Logical. Specify if the contrast function has been used after the emmeans function

contrast.name
Character. Corresponds to the label of the column in which the contrasts are specified Default value is 'contrast'.

stat.name
Character. Indicates the name of the variable that report the statistics in report.quanti function. Default = "Statistics"

Value

A desc object

Details

No detail.

See also

Examples



# NOT RUN {
x=desc()

# }