'LS Means' statistics reporting

Creates a desc object for "LS Means" statistics reporting.

report.lsmeans(lsm, x1 = "treatment", x2 = NULL, x3 = NULL, data,
  variable.name = "Statistics", at.row = NULL, infer = c(T, T),
  type = "link", contrast = F, contrast.name = "contrast",
  x1.name = "treatment", x2.name = NULL, x3.name = NULL)

Arguments

lsm
emmGrid object (result of a emmeans call)

x1
Character Mandatory. Indicating a factor in the data (can be an intercept: see example). Levels will be displayed in columns

x2
Character indicating a factor in the data. Levels will be displayed in rows

x3
Character indicating a factor in the data. Levels will be displayed in rows

data
Data.frame object from which the Least Square means are coming from

variable.name
Character. The label of the column which indicates the statistics reported.

at.row
Character. Passed to spacetable function. Used to space the results per levels of the mentioned variable

infer
A vector of one or two logical values. Passed to summary.emmGrid function.

type
Character. Type of prediction desired. Passed to summary.emmGrid function. Can be "link" or "response"

contrast
Logical. Specify if the contrast function has been used after the emmeans function (see examples)

contrast.name
Character. Corresponds to the label of the column in which the contrasts are specified (see example). Default value is 'contrast'.

x1.name
Character. Deprecated (replaced by x1)

x2.name
Character. Deprecated (replaced by x2)

x3.name
Character Deprecated (replaced by x3)

Value

A desc object that can be used by the report.doc function.

Details

You can produce formatted Least Square Means table for up to 3 factors. It doesn't work for quantitative covariates.

See examples below.

See also

Examples


library(emmeans)
library(lme4)

data(data)

#Simple lm model

mod=lm(Petal.Width~Species,data=iris)
raw.lsm=emmeans(mod,~Species)
report.lsmeans(raw.lsm,"Species",data=iris)

#>      Statistics      setosa  versicolor   virginica
#> 1 Estimate (SE)  0.25(0.03)  1.33(0.03)  2.03(0.03)
#> 2        95% CI [0.19;0.30] [1.27;1.38] [1.97;2.08]
#> 3       P-value      <0.001      <0.001      <0.001


# In case of just one intercept you must use a workaround...
iris$int=1
mod=glm(Species~int,data=iris,family=binomial)
raw.lsm=emmeans(mod,~int)
iris$int=as.factor(iris$int)
report.lsmeans(raw.lsm,"int",data=iris)

#>      Statistics           1
#> 1 Estimate (SE)  0.69(0.17)
#> 2        95% CI [0.35;1.03]
#> 3       P-value      <0.001


#Mixed model example using lme4

mod=lmer(y_numeric~GROUP+TIMEPOINT+GROUP*TIMEPOINT+(1|SUBJID),data=data)

#> singular fit

raw.lsm=emmeans(mod,~GROUP|TIMEPOINT)
report.lsmeans(lsm=raw.lsm,x1="GROUP",x2="TIMEPOINT",at.row="TIMEPOINT",data=data)

#>    TIMEPOINT    Statistics             A             B             C
#> 1         D0 Estimate (SE)   -0.93(0.20)   -0.67(0.24)   -1.19(0.26)
#> 2         D0        95% CI [-1.32;-0.54] [-1.16;-0.19] [-1.72;-0.66]
#> 3         D0       P-value        <0.001         0.010        <0.001
#> 4                                                                   
#> 5         D1 Estimate (SE)    1.83(0.20)    4.17(0.24)    4.98(0.26)
#> 6         D1        95% CI   [1.44;2.22]   [3.70;4.63]   [4.46;5.50]
#> 7         D1       P-value        <0.001        <0.001        <0.001
#> 8                                                                   
#> 9         D2 Estimate (SE)    1.97(0.20)    4.04(0.24)    4.90(0.26)
#> 10        D2        95% CI   [1.58;2.35]   [3.56;4.52]   [4.38;5.42]
#> 11        D2       P-value        <0.001        <0.001        <0.001
#> 12                                                                  
#> 13        D3 Estimate (SE)    1.78(0.19)    3.81(0.24)    5.07(0.27)
#> 14        D3        95% CI   [1.40;2.16]   [3.33;4.29]   [4.54;5.61]
#> 15        D3       P-value        <0.001        <0.001        <0.001
#> 16                                                                  
#> 17        D4 Estimate (SE)    1.83(0.20)    3.80(0.24)    5.17(0.27)
#> 18        D4        95% CI   [1.44;2.22]   [3.32;4.28]   [4.64;5.71]
#> 19        D4       P-value        <0.001        <0.001        <0.001
#> 20                                                                  
#> 21        D5 Estimate (SE)    2.27(0.19)    3.64(0.24)    4.43(0.26)
#> 22        D5        95% CI   [1.89;2.65]   [3.18;4.11]   [3.91;4.95]
#> 23        D5       P-value        <0.001        <0.001        <0.001



# GLM model with specific contrast

warp.lm <- lm(breaks ~ wool+tension+wool:tension, data = warpbreaks)
warp.emm <- emmeans(warp.lm, ~ tension | wool)
contr=contrast(warp.emm, "trt.vs.ctrl", ref = "M")
report.lsmeans(lsm=contr,x1="wool",data=warpbreaks,contrast=TRUE,at.row="contrast")

#>   contrast    Statistics              A              B
#> 1    L - M Estimate (SE)    20.56(5.16)    -0.56(5.16)
#> 2    L - M        95% CI   [8.74;32.37] [-12.37;11.26]
#> 3    L - M       P-value         <0.001          0.990
#> 4                                                     
#> 5    H - M Estimate (SE)     0.56(5.16)   -10.00(5.16)
#> 6    H - M        95% CI [-11.26;12.37]  [-21.81;1.81]
#> 7    H - M       P-value          0.990          0.110