Descriptive "Qualitative" statistics (frequencies and percentages) reporting

Compute and report frequencies and percentages by levels of y (in rows) and by levels of x1 (in columns) and x2 in rows.

report.quali(data, y = NULL, x1 = NULL, x2 = NULL, y.label = y,
  x2.label = NULL, y.levels.label = "Levels", total = F, round = 2,
  at.row = NULL, percent.col = T, subjid = NULL, remove.zero = F,
  drop.y = NULL, drop.x1 = NULL, drop.x2 = NULL,
  remove.missing = F)

Arguments

data
a data.frame object

y
Character indicating a factor in the data (the response)

x1
Character indicating a factor in the data (levels will be displayed in columns)

x2
Character indicating a factor in the data (levels will be displayed in rows). Only possible if x1 is not NULL.

y.label
Character indicating the label for y to be displayed in the title of the table

x2.label
Character indicating the label for x2 parameter, only if x2 is not null

y.levels.label
Character. Indicates the label of the column in which the levels of y are displayed

total
Logical Indicates if a column Total should be added or not

round
Integer Indicates the number of digits to round percentages

at.row
Character. Pass to spacetable function. Used to space the results (see example below)

percent.col
Logical By default it is set to T to indicate that column percentages should be reported. If set to False, row percentages are reported.

subjid
Character Indicates in the data.frame the name of the column used to identify the Id of the subjects. If not null, it adds in the headers the number of unique subject per levels of x1 or y (if x1 is null).

remove.zero
Logical. Remove the levels in the contingency table for which there is no observation.

drop.y
Character. Indicates one or several levels of the y factor that you want to drop in the result

drop.x1
Character. Indicates one or several levels of the x1 factor that you want to drop in the result

drop.x2
Character. Indicates one or several levels of the x2 factor that you want to drop in the result

remove.missing
Logical. default to TRUE. If TRUE number of missing values are reported and percentages take into account the number of missing value in the calculation. If set to FALSE, the missing values regarding the response factor y are ignored and percentages are computed on non missing values only.

Value

A desc object

Details

This function computes and reports qualitative statistics by level of y and by level of x1 (if not null) and x2 (if not null).

See examples to show the results. If total=T, the last column is the statistics performed overall levels of the explicative variables x1.

Note that missing values are counted in the calculation of the percentages.

See also

Examples



data(data)

# No explicative variable changing y labels (option y.levels.label)
report.quali(data=data,y="y_logistic",
y.label="Clinical cure",y.levels.label="Levels")

#> 
#> ############################################
#> Qualitative descriptive statistics of : Clinical cure
#> ############################################
#> 
#>   Levels    Statistics Clinical cure
#> 1      0  n (column %)   186(46.97%)
#> 2      1  n (column %)   200(50.51%)
#> 3         Missing n(%)     10(2.53%)
#> 
#> ############################################
#> 


# 1 explicative variable (option x1:  levels in columns)
report.quali(data=data,y="y_logistic",x1="GROUP",y.levels.label="Clinical cure")

#> 
#> ############################################
#> Qualitative descriptive statistics of : y_logistic
#> ############################################
#> 
#>   Clinical cure    Statistics          A          B          C
#> 1             0  n (column %) 79(43.89%) 60(50.00%) 47(48.96%)
#> 2             1  n (column %) 97(53.89%) 59(49.17%) 44(45.83%)
#> 3                Missing n(%)   4(2.22%)   1(0.83%)   5(5.21%)
#> 
#> ############################################
#> 


# 2 explicative variables (x1, levels in columns, option x2, levels in rows),
# spcacing the results (option at.row)
report.quali(data=data,y="y_logistic",x1="GROUP",
x2="TIMEPOINT",y.levels.label="Clinical cure",x2.label="Days",at.row="Days")

#> 
#> ############################################
#> Qualitative descriptive statistics of : y_logistic
#> ############################################
#> 
#>    Days Clinical cure    Statistics          A          B          C
#> 1    D0             0  n (column %) 11(36.67%) 11(55.00%)  7(43.75%)
#> 2    D0             1  n (column %) 18(60.00%)  8(40.00%)  7(43.75%)
#> 3    D0                Missing n(%)   1(3.33%)   1(5.00%)  2(12.50%)
#> 4                                                                   
#> 5    D1             0  n (column %)  7(23.33%) 13(65.00%)  8(50.00%)
#> 6    D1             1  n (column %) 21(70.00%)  7(35.00%)  7(43.75%)
#> 7    D1                Missing n(%)   2(6.67%)      0(0%)   1(6.25%)
#> 8                                                                   
#> 9    D2             0  n (column %) 18(60.00%)  7(35.00%) 11(68.75%)
#> 10   D2             1  n (column %) 12(40.00%) 13(65.00%)  5(31.25%)
#> 11   D2                Missing n(%)      0(0%)      0(0%)      0(0%)
#> 12                                                                  
#> 13   D3             0  n (column %) 11(36.67%) 10(50.00%)  7(43.75%)
#> 14   D3             1  n (column %) 19(63.33%) 10(50.00%)  9(56.25%)
#> 15   D3                Missing n(%)      0(0%)      0(0%)      0(0%)
#> 16                                                                  
#> 17   D4             0  n (column %) 18(60.00%) 12(60.00%)  6(37.50%)
#> 18   D4             1  n (column %) 12(40.00%)  8(40.00%)  8(50.00%)
#> 19   D4                Missing n(%)      0(0%)      0(0%)  2(12.50%)
#> 20                                                                  
#> 21   D5             0  n (column %) 14(46.67%)  7(35.00%)  8(50.00%)
#> 22   D5             1  n (column %) 15(50.00%) 13(65.00%)  8(50.00%)
#> 23   D5                Missing n(%)   1(3.33%)      0(0%)      0(0%)
#> 
#> ############################################
#> 


# 2 explicative variables with row percentages (option percent.col=F)
report.quali(data=data,y="y_logistic",x1="GROUP",
x2="TIMEPOINT",percent.col=FALSE,x2.label="Days",y.levels.label="Clinical cure")

#> 
#> ############################################
#> Qualitative descriptive statistics of : y_logistic
#> ############################################
#> 
#>    Days Clinical cure    Statistics          A          B          C
#> 1    D0             0     n (row %) 11(37.93%) 11(37.93%)  7(24.14%)
#> 2    D0             1     n (row %) 18(54.55%)  8(24.24%)  7(21.21%)
#> 3    D0                Missing n(%)  1(25.00%)  1(25.00%)  2(50.00%)
#> 4                                                                   
#> 5    D1             0     n (row %)  7(25.00%) 13(46.43%)  8(28.57%)
#> 6    D1             1     n (row %) 21(60.00%)  7(20.00%)  7(20.00%)
#> 7    D1                Missing n(%)  2(66.67%)      0(0%)  1(33.33%)
#> 8                                                                   
#> 9    D2             0     n (row %) 18(50.00%)  7(19.44%) 11(30.56%)
#> 10   D2             1     n (row %) 12(40.00%) 13(43.33%)  5(16.67%)
#> 11   D2                Missing n(%)          0          0          0
#> 12                                                                  
#> 13   D3             0     n (row %) 11(39.29%) 10(35.71%)  7(25.00%)
#> 14   D3             1     n (row %) 19(50.00%) 10(26.32%)  9(23.68%)
#> 15   D3                Missing n(%)          0          0          0
#> 16                                                                  
#> 17   D4             0     n (row %) 18(50.00%) 12(33.33%)  6(16.67%)
#> 18   D4             1     n (row %) 12(42.86%)  8(28.57%)  8(28.57%)
#> 19   D4                Missing n(%)      0(0%)      0(0%) 2(100.00%)
#> 20                                                                  
#> 21   D5             0     n (row %) 14(48.28%)  7(24.14%)  8(27.59%)
#> 22   D5             1     n (row %) 15(41.67%) 13(36.11%)  8(22.22%)
#> 23   D5                Missing n(%) 1(100.00%)      0(0%)      0(0%)
#> 
#> ############################################
#> 


# Add Total number of subjects in headers (option subjid="SUBJID")
report.quali(data=data,y="y_logistic",x1="GROUP",
x2="TIMEPOINT",at.row="TIMEPOINT",subjid="SUBJID")

#> 
#> ############################################
#> Qualitative descriptive statistics of : y_logistic
#> ############################################
#> 
#>    TIMEPOINT Levels    Statistics   A (N=30)   B (N=21)   C (N=17)
#> 1         D0      0  n (column %) 11(36.67%) 11(55.00%)  7(43.75%)
#> 2         D0      1  n (column %) 18(60.00%)  8(40.00%)  7(43.75%)
#> 3         D0         Missing n(%)   1(3.33%)   1(5.00%)  2(12.50%)
#> 4                                                                 
#> 5         D1      0  n (column %)  7(23.33%) 13(65.00%)  8(50.00%)
#> 6         D1      1  n (column %) 21(70.00%)  7(35.00%)  7(43.75%)
#> 7         D1         Missing n(%)   2(6.67%)      0(0%)   1(6.25%)
#> 8                                                                 
#> 9         D2      0  n (column %) 18(60.00%)  7(35.00%) 11(68.75%)
#> 10        D2      1  n (column %) 12(40.00%) 13(65.00%)  5(31.25%)
#> 11        D2         Missing n(%)      0(0%)      0(0%)      0(0%)
#> 12                                                                
#> 13        D3      0  n (column %) 11(36.67%) 10(50.00%)  7(43.75%)
#> 14        D3      1  n (column %) 19(63.33%) 10(50.00%)  9(56.25%)
#> 15        D3         Missing n(%)      0(0%)      0(0%)      0(0%)
#> 16                                                                
#> 17        D4      0  n (column %) 18(60.00%) 12(60.00%)  6(37.50%)
#> 18        D4      1  n (column %) 12(40.00%)  8(40.00%)  8(50.00%)
#> 19        D4         Missing n(%)      0(0%)      0(0%)  2(12.50%)
#> 20                                                                
#> 21        D5      0  n (column %) 14(46.67%)  7(35.00%)  8(50.00%)
#> 22        D5      1  n (column %) 15(50.00%) 13(65.00%)  8(50.00%)
#> 23        D5         Missing n(%)   1(3.33%)      0(0%)      0(0%)
#> 
#> ############################################
#> 


# Round percentages to 1 digit (option round=1)
report.quali(data=data,y="y_logistic",x1="GROUP",
x2="TIMEPOINT",at.row="TIMEPOINT",subjid="SUBJID",round=1)

#> 
#> ############################################
#> Qualitative descriptive statistics of : y_logistic
#> ############################################
#> 
#>    TIMEPOINT Levels    Statistics  A (N=30)  B (N=21)  C (N=17)
#> 1         D0      0  n (column %) 11(36.7%) 11(55.0%)  7(43.8%)
#> 2         D0      1  n (column %) 18(60.0%)  8(40.0%)  7(43.8%)
#> 3         D0         Missing n(%)   1(3.3%)   1(5.0%)  2(12.5%)
#> 4                                                              
#> 5         D1      0  n (column %)  7(23.3%) 13(65.0%)  8(50.0%)
#> 6         D1      1  n (column %) 21(70.0%)  7(35.0%)  7(43.8%)
#> 7         D1         Missing n(%)   2(6.7%)   0(0.0%)   1(6.2%)
#> 8                                                              
#> 9         D2      0  n (column %) 18(60.0%)  7(35.0%) 11(68.8%)
#> 10        D2      1  n (column %) 12(40.0%) 13(65.0%)  5(31.2%)
#> 11        D2         Missing n(%)   0(0.0%)   0(0.0%)   0(0.0%)
#> 12                                                             
#> 13        D3      0  n (column %) 11(36.7%) 10(50.0%)  7(43.8%)
#> 14        D3      1  n (column %) 19(63.3%) 10(50.0%)  9(56.2%)
#> 15        D3         Missing n(%)   0(0.0%)   0(0.0%)   0(0.0%)
#> 16                                                             
#> 17        D4      0  n (column %) 18(60.0%) 12(60.0%)  6(37.5%)
#> 18        D4      1  n (column %) 12(40.0%)  8(40.0%)  8(50.0%)
#> 19        D4         Missing n(%)   0(0.0%)   0(0.0%)  2(12.5%)
#> 20                                                             
#> 21        D5      0  n (column %) 14(46.7%)  7(35.0%)  8(50.0%)
#> 22        D5      1  n (column %) 15(50.0%) 13(65.0%)  8(50.0%)
#> 23        D5         Missing n(%)   1(3.3%)   0(0.0%)   0(0.0%)
#> 
#> ############################################
#> 



# Qualitative statistics with a response with more than one levels
tab=report.quali(data=data,y="y_poisson",x1="GROUP",
x2="TIMEPOINT",at.row="TIMEPOINT",subjid="SUBJID",round=1)

#> data[,'y_poisson']has been transformed into a factor


# Print formatted results
tab

#> 
#> ############################################
#> Qualitative descriptive statistics of : y_poisson
#> ############################################
#> 
#>    TIMEPOINT Levels    Statistics  A (N=30)  B (N=21) C (N=17)
#> 1         D0      1  n (column %)   0(0.0%)   0(0.0%)  0(0.0%)
#> 2         D0      2  n (column %)   0(0.0%)   0(0.0%)  1(6.2%)
#> 3         D0      3  n (column %)  9(30.0%)  4(20.0%) 4(25.0%)
#> 4         D0      4  n (column %) 10(33.3%)  8(40.0%) 8(50.0%)
#> 5         D0      5  n (column %) 10(33.3%)  7(35.0%) 2(12.5%)
#> 6         D0      6  n (column %)   0(0.0%)   0(0.0%)  1(6.2%)
#> 7         D0      7  n (column %)   0(0.0%)   1(5.0%)  0(0.0%)
#> 8         D0      8  n (column %)   0(0.0%)   0(0.0%)  0(0.0%)
#> 9         D0      9  n (column %)   0(0.0%)   0(0.0%)  0(0.0%)
#> 10        D0     10  n (column %)   0(0.0%)   0(0.0%)  0(0.0%)
#> 11        D0     11  n (column %)   0(0.0%)   0(0.0%)  0(0.0%)
#> 12        D0         Missing n(%)   1(3.3%)   0(0.0%)  0(0.0%)
#> 13                                                            
#> 14        D1      1  n (column %)   2(6.7%)   0(0.0%)  0(0.0%)
#> 15        D1      2  n (column %)  9(30.0%)   0(0.0%)  0(0.0%)
#> 16        D1      3  n (column %)  9(30.0%)   0(0.0%)  0(0.0%)
#> 17        D1      4  n (column %)  7(23.3%)   0(0.0%) 4(25.0%)
#> 18        D1      5  n (column %)   1(3.3%)   0(0.0%) 8(50.0%)
#> 19        D1      6  n (column %)   1(3.3%)   0(0.0%)  1(6.2%)
#> 20        D1      7  n (column %)   0(0.0%)   0(0.0%)  1(6.2%)
#> 21        D1      8  n (column %)   0(0.0%) 10(50.0%)  0(0.0%)
#> 22        D1      9  n (column %)   0(0.0%)  6(30.0%)  0(0.0%)
#> 23        D1     10  n (column %)   0(0.0%)  4(20.0%)  0(0.0%)
#> 24        D1     11  n (column %)   0(0.0%)   0(0.0%)  0(0.0%)
#> 25        D1         Missing n(%)   1(3.3%)   0(0.0%) 2(12.5%)
#> 26                                                            
#> 27        D2      1  n (column %)   0(0.0%)   0(0.0%)  0(0.0%)
#> 28        D2      2  n (column %)  9(30.0%)   0(0.0%)  0(0.0%)
#> 29        D2      3  n (column %) 10(33.3%)   0(0.0%) 2(12.5%)
#> 30        D2      4  n (column %)  7(23.3%)   0(0.0%) 4(25.0%)
#> 31        D2      5  n (column %)  3(10.0%)   0(0.0%) 6(37.5%)
#> 32        D2      6  n (column %)   0(0.0%)   0(0.0%) 2(12.5%)
#> 33        D2      7  n (column %)   0(0.0%)   1(5.0%)  1(6.2%)
#> 34        D2      8  n (column %)   0(0.0%)  4(20.0%)  1(6.2%)
#> 35        D2      9  n (column %)   0(0.0%)  9(45.0%)  0(0.0%)
#> 36        D2     10  n (column %)   0(0.0%)  4(20.0%)  0(0.0%)
#> 37        D2     11  n (column %)   0(0.0%)  2(10.0%)  0(0.0%)
#> 38        D2         Missing n(%)   1(3.3%)   0(0.0%)  0(0.0%)
#> 39                                                            
#> 40        D3      1  n (column %)  4(13.3%)   0(0.0%)  0(0.0%)
#> 41        D3      2  n (column %)  6(20.0%)   0(0.0%)  0(0.0%)
#> 42        D3      3  n (column %) 10(33.3%)   0(0.0%) 2(12.5%)
#> 43        D3      4  n (column %)  9(30.0%)   0(0.0%) 2(12.5%)
#> 44        D3      5  n (column %)   1(3.3%)   0(0.0%) 9(56.2%)
#> 45        D3      6  n (column %)   0(0.0%)   0(0.0%) 2(12.5%)
#> 46        D3      7  n (column %)   0(0.0%)   0(0.0%)  0(0.0%)
#> 47        D3      8  n (column %)   0(0.0%)  5(25.0%)  1(6.2%)
#> 48        D3      9  n (column %)   0(0.0%) 10(50.0%)  0(0.0%)
#> 49        D3     10  n (column %)   0(0.0%)  3(15.0%)  0(0.0%)
#> 50        D3     11  n (column %)   0(0.0%)  2(10.0%)  0(0.0%)
#> 51        D3         Missing n(%)   0(0.0%)   0(0.0%)  0(0.0%)
#> 52                                                            
#> 53        D4      1  n (column %)   2(6.7%)   0(0.0%)  0(0.0%)
#> 54        D4      2  n (column %)  6(20.0%)   0(0.0%)  0(0.0%)
#> 55        D4      3  n (column %) 12(40.0%)   0(0.0%)  1(6.2%)
#> 56        D4      4  n (column %)  7(23.3%)   0(0.0%) 4(25.0%)
#> 57        D4      5  n (column %)  3(10.0%)   0(0.0%) 4(25.0%)
#> 58        D4      6  n (column %)   0(0.0%)   0(0.0%) 4(25.0%)
#> 59        D4      7  n (column %)   0(0.0%)  2(10.0%)  1(6.2%)
#> 60        D4      8  n (column %)   0(0.0%)   1(5.0%)  0(0.0%)
#> 61        D4      9  n (column %)   0(0.0%) 12(60.0%)  0(0.0%)
#> 62        D4     10  n (column %)   0(0.0%)  4(20.0%)  0(0.0%)
#> 63        D4     11  n (column %)   0(0.0%)   0(0.0%)  0(0.0%)
#> 64        D4         Missing n(%)   0(0.0%)   1(5.0%) 2(12.5%)
#> 65                                                            
#> 66        D5      1  n (column %)   0(0.0%)   0(0.0%)  0(0.0%)
#> 67        D5      2  n (column %)  6(20.0%)   0(0.0%)  0(0.0%)
#> 68        D5      3  n (column %) 14(46.7%)   0(0.0%)  0(0.0%)
#> 69        D5      4  n (column %)  4(13.3%)   0(0.0%) 8(50.0%)
#> 70        D5      5  n (column %)   2(6.7%)   0(0.0%) 5(31.2%)
#> 71        D5      6  n (column %)   2(6.7%)   0(0.0%) 2(12.5%)
#> 72        D5      7  n (column %)   0(0.0%)  2(10.0%)  1(6.2%)
#> 73        D5      8  n (column %)   0(0.0%)  9(45.0%)  0(0.0%)
#> 74        D5      9  n (column %)   0(0.0%)  5(25.0%)  0(0.0%)
#> 75        D5     10  n (column %)   0(0.0%)  4(20.0%)  0(0.0%)
#> 76        D5     11  n (column %)   0(0.0%)   0(0.0%)  0(0.0%)
#> 77        D5         Missing n(%)   2(6.7%)   0(0.0%)  0(0.0%)
#> 
#> ############################################
#> 



#Getting raw output (unformatted) 
tab$raw.output

#>     GROUP TIMEPOINT y_poisson Freq.x Freq.y
#> 1       A        D0         1      0     30
#> 2       A        D0         2      0     30
#> 3       A        D0         3      9     30
#> 4       A        D0         4     10     30
#> 5       A        D0         5     10     30
#> 6       A        D0         6      0     30
#> 7       A        D0         7      0     30
#> 8       A        D0         8      0     30
#> 9       A        D0         9      0     30
#> 10      A        D0        10      0     30
#> 11      A        D0        11      0     30
#> 12      A        D0      <NA>      1     30
#> 13      A        D1         1      2     30
#> 14      A        D1         2      9     30
#> 15      A        D1         3      9     30
#> 16      A        D1         4      7     30
#> 17      A        D1         5      1     30
#> 18      A        D1         6      1     30
#> 19      A        D1         7      0     30
#> 20      A        D1         8      0     30
#> 21      A        D1         9      0     30
#> 22      A        D1        10      0     30
#> 23      A        D1        11      0     30
#> 24      A        D1      <NA>      1     30
#> 25      A        D2         1      0     30
#> 26      A        D2         2      9     30
#> 27      A        D2         3     10     30
#> 28      A        D2         4      7     30
#> 29      A        D2         5      3     30
#> 30      A        D2         6      0     30
#> 31      A        D2         7      0     30
#> 32      A        D2         8      0     30
#> 33      A        D2         9      0     30
#> 34      A        D2        10      0     30
#> 35      A        D2        11      0     30
#> 36      A        D2      <NA>      1     30
#> 37      A        D3         1      4     30
#> 38      A        D3         2      6     30
#> 39      A        D3         3     10     30
#> 40      A        D3         4      9     30
#> 41      A        D3         5      1     30
#> 42      A        D3         6      0     30
#> 43      A        D3         7      0     30
#> 44      A        D3         8      0     30
#> 45      A        D3         9      0     30
#> 46      A        D3        10      0     30
#> 47      A        D3        11      0     30
#> 48      A        D3      <NA>      0     30
#> 49      A        D4         1      2     30
#> 50      A        D4         2      6     30
#> 51      A        D4         3     12     30
#> 52      A        D4         4      7     30
#> 53      A        D4         5      3     30
#> 54      A        D4         6      0     30
#> 55      A        D4         7      0     30
#> 56      A        D4         8      0     30
#> 57      A        D4         9      0     30
#> 58      A        D4        10      0     30
#> 59      A        D4        11      0     30
#> 60      A        D4      <NA>      0     30
#> 61      A        D5         1      0     30
#> 62      A        D5         2      6     30
#> 63      A        D5         3     14     30
#> 64      A        D5         4      4     30
#> 65      A        D5         5      2     30
#> 66      A        D5         6      2     30
#> 67      A        D5         7      0     30
#> 68      A        D5         8      0     30
#> 69      A        D5         9      0     30
#> 70      A        D5        10      0     30
#> 71      A        D5        11      0     30
#> 72      A        D5      <NA>      2     30
#> 73      B        D0         1      0     20
#> 74      B        D0         2      0     20
#> 75      B        D0         3      4     20
#> 76      B        D0         4      8     20
#> 77      B        D0         5      7     20
#> 78      B        D0         6      0     20
#> 79      B        D0         7      1     20
#> 80      B        D0         8      0     20
#> 81      B        D0         9      0     20
#> 82      B        D0        10      0     20
#> 83      B        D0        11      0     20
#> 84      B        D0      <NA>      0     20
#> 85      B        D1         1      0     20
#> 86      B        D1         2      0     20
#> 87      B        D1         3      0     20
#> 88      B        D1         4      0     20
#> 89      B        D1         5      0     20
#> 90      B        D1         6      0     20
#> 91      B        D1         7      0     20
#> 92      B        D1         8     10     20
#> 93      B        D1         9      6     20
#> 94      B        D1        10      4     20
#> 95      B        D1        11      0     20
#> 96      B        D1      <NA>      0     20
#> 97      B        D2         1      0     20
#> 98      B        D2         2      0     20
#> 99      B        D2         3      0     20
#> 100     B        D2         4      0     20
#> 101     B        D2         5      0     20
#> 102     B        D2         6      0     20
#> 103     B        D2         7      1     20
#> 104     B        D2         8      4     20
#> 105     B        D2         9      9     20
#> 106     B        D2        10      4     20
#> 107     B        D2        11      2     20
#> 108     B        D2      <NA>      0     20
#> 109     B        D3         1      0     20
#> 110     B        D3         2      0     20
#> 111     B        D3         3      0     20
#> 112     B        D3         4      0     20
#> 113     B        D3         5      0     20
#> 114     B        D3         6      0     20
#> 115     B        D3         7      0     20
#> 116     B        D3         8      5     20
#> 117     B        D3         9     10     20
#> 118     B        D3        10      3     20
#> 119     B        D3        11      2     20
#> 120     B        D3      <NA>      0     20
#> 121     B        D4         1      0     20
#> 122     B        D4         2      0     20
#> 123     B        D4         3      0     20
#> 124     B        D4         4      0     20
#> 125     B        D4         5      0     20
#> 126     B        D4         6      0     20
#> 127     B        D4         7      2     20
#> 128     B        D4         8      1     20
#> 129     B        D4         9     12     20
#> 130     B        D4        10      4     20
#> 131     B        D4        11      0     20
#> 132     B        D4      <NA>      1     20
#> 133     B        D5         1      0     20
#> 134     B        D5         2      0     20
#> 135     B        D5         3      0     20
#> 136     B        D5         4      0     20
#> 137     B        D5         5      0     20
#> 138     B        D5         6      0     20
#> 139     B        D5         7      2     20
#> 140     B        D5         8      9     20
#> 141     B        D5         9      5     20
#> 142     B        D5        10      4     20
#> 143     B        D5        11      0     20
#> 144     B        D5      <NA>      0     20
#> 145     C        D0         1      0     16
#> 146     C        D0         2      1     16
#> 147     C        D0         3      4     16
#> 148     C        D0         4      8     16
#> 149     C        D0         5      2     16
#> 150     C        D0         6      1     16
#> 151     C        D0         7      0     16
#> 152     C        D0         8      0     16
#> 153     C        D0         9      0     16
#> 154     C        D0        10      0     16
#> 155     C        D0        11      0     16
#> 156     C        D0      <NA>      0     16
#> 157     C        D1         1      0     16
#> 158     C        D1         2      0     16
#> 159     C        D1         3      0     16
#> 160     C        D1         4      4     16
#> 161     C        D1         5      8     16
#> 162     C        D1         6      1     16
#> 163     C        D1         7      1     16
#> 164     C        D1         8      0     16
#> 165     C        D1         9      0     16
#> 166     C        D1        10      0     16
#> 167     C        D1        11      0     16
#> 168     C        D1      <NA>      2     16
#> 169     C        D2         1      0     16
#> 170     C        D2         2      0     16
#> 171     C        D2         3      2     16
#> 172     C        D2         4      4     16
#> 173     C        D2         5      6     16
#> 174     C        D2         6      2     16
#> 175     C        D2         7      1     16
#> 176     C        D2         8      1     16
#> 177     C        D2         9      0     16
#> 178     C        D2        10      0     16
#> 179     C        D2        11      0     16
#> 180     C        D2      <NA>      0     16
#> 181     C        D3         1      0     16
#> 182     C        D3         2      0     16
#> 183     C        D3         3      2     16
#> 184     C        D3         4      2     16
#> 185     C        D3         5      9     16
#> 186     C        D3         6      2     16
#> 187     C        D3         7      0     16
#> 188     C        D3         8      1     16
#> 189     C        D3         9      0     16
#> 190     C        D3        10      0     16
#> 191     C        D3        11      0     16
#> 192     C        D3      <NA>      0     16
#> 193     C        D4         1      0     16
#> 194     C        D4         2      0     16
#> 195     C        D4         3      1     16
#> 196     C        D4         4      4     16
#> 197     C        D4         5      4     16
#> 198     C        D4         6      4     16
#> 199     C        D4         7      1     16
#> 200     C        D4         8      0     16
#> 201     C        D4         9      0     16
#> 202     C        D4        10      0     16
#> 203     C        D4        11      0     16
#> 204     C        D4      <NA>      2     16
#> 205     C        D5         1      0     16
#> 206     C        D5         2      0     16
#> 207     C        D5         3      0     16
#> 208     C        D5         4      8     16
#> 209     C        D5         5      5     16
#> 210     C        D5         6      2     16
#> 211     C        D5         7      1     16
#> 212     C        D5         8      0     16
#> 213     C        D5         9      0     16
#> 214     C        D5        10      0     16
#> 215     C        D5        11      0     16
#> 216     C        D5      <NA>      0     16