'Hierarchical Qualitative' statistics reporting (experimental)

This function is mainly used to compute qualitative statistics when there are several events per statistical unit. Often used for reporting adverse events, medical history or concomitant treatments.

It reports frequencies and percentages according to hierarchical levels of two factors.

Typically, adverse event are classified according to System Organ Class (SOC) and then sub classified by Prefered Terms (PT). Several observations of a same adverse event can be observed several times on the same subject. It's then useful to know how many persons are concerned by at least one of those adverse events and report the frequencies for each classifications: SOC and PT.

This is exactly what this function does.

For more examples see the website: ClinReport website

report.quali.hlev(data, subjid = NULL, x1 = NULL, var_upper, var_lower,
  lower.levels = "Lower.Levels", upper.levels = "Upper.Levels",
  x1.label = NULL)

Arguments

data
A data frame

subjid
A character

x1
A character. Indicates a factor in the data frame.

var_upper
A character. Indicates a factor in the data frame which corresponds to the factor with the higher number of levels (typically SOC variable)

var_lower
A character. Indicates a factor in the data frame which corresponds to the factor with the lower number of levels (typically PT variable)

lower.levels
A character. The label to be displayed in the table for the lower terms

upper.levels
A character. The label to be displayed in the table for the upper terms

upper.levels
A character. Not used for now

Value

A desc object that can be used by the report.doc function.

Details

The subjid argument is mandatory for this function.

See also

Examples


data(adverse_event)

test=report.quali.hlev(data=adverse_event,subjid="SUBJID",var_upper="PTNAME",
var_lower="SOCNAME",lower.levels="System Organ Class",upper.levels="Prefered Terms",x1="randtrt")

# show results in console
test

#> 
#> ############################################
#> Hierarchichal
#> ############################################
#> 
#>                       System Organ Class n (%) SOCNAME At least one SOCNAME
#> 1                                    ALL     451(100%)     195/321 = 60.75%
#> 2                                                                          
#> 3                  BEHAVIOURAL DISORDERS      3(0.67%)        1/321 = 0.31%
#> 4                                                                          
#> 5   BLOOD AND LYMPHATIC SYSTEM DISORDERS     10(2.22%)        7/321 = 2.18%
#> 6   BLOOD AND LYMPHATIC SYSTEM DISORDERS                                   
#> 7   BLOOD AND LYMPHATIC SYSTEM DISORDERS                                   
#> 8   BLOOD AND LYMPHATIC SYSTEM DISORDERS                                   
#> 9   BLOOD AND LYMPHATIC SYSTEM DISORDERS                                   
#> 10                                                                         
#> 11      CARDIO-VASCULAR SYSTEM DISORDERS     21(4.66%)       18/321 = 5.61%
#> 12      CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 13      CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 14      CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 15      CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 16      CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 17      CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 18      CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 19                                                                         
#> 20             DIGESTIVE TRACT DISORDERS    72(15.96%)      55/321 = 17.13%
#> 21             DIGESTIVE TRACT DISORDERS                                   
#> 22             DIGESTIVE TRACT DISORDERS                                   
#> 23             DIGESTIVE TRACT DISORDERS                                   
#> 24             DIGESTIVE TRACT DISORDERS                                   
#> 25             DIGESTIVE TRACT DISORDERS                                   
#> 26             DIGESTIVE TRACT DISORDERS                                   
#> 27             DIGESTIVE TRACT DISORDERS                                   
#> 28             DIGESTIVE TRACT DISORDERS                                   
#> 29             DIGESTIVE TRACT DISORDERS                                   
#> 30             DIGESTIVE TRACT DISORDERS                                   
#> 31             DIGESTIVE TRACT DISORDERS                                   
#> 32             DIGESTIVE TRACT DISORDERS                                   
#> 33             DIGESTIVE TRACT DISORDERS                                   
#> 34             DIGESTIVE TRACT DISORDERS                                   
#> 35                                                                         
#> 36           EAR AND LABYRINTH DISORDERS      3(0.67%)        3/321 = 0.93%
#> 37           EAR AND LABYRINTH DISORDERS                                   
#> 38                                                                         
#> 39                         EYE DISORDERS     10(2.22%)       10/321 = 3.12%
#> 40                         EYE DISORDERS                                   
#> 41                         EYE DISORDERS                                   
#> 42                         EYE DISORDERS                                   
#> 43                                                                         
#> 44              HEPATO-BILIARY DISORDERS      5(1.11%)        4/321 = 1.25%
#> 45                                                                         
#> 46               MAMMARY GLAND DISORDERS      1(0.22%)        1/321 = 0.31%
#> 47                                                                         
#> 48    METABOLISM AND NUTRITION DISORDERS      7(1.55%)        6/321 = 1.87%
#> 49    METABOLISM AND NUTRITION DISORDERS                                   
#> 50                                                                         
#> 51             MUSCULOSKELETAL DISORDERS     15(3.33%)       12/321 = 3.74%
#> 52             MUSCULOSKELETAL DISORDERS                                   
#> 53             MUSCULOSKELETAL DISORDERS                                   
#> 54             MUSCULOSKELETAL DISORDERS                                   
#> 55             MUSCULOSKELETAL DISORDERS                                   
#> 56             MUSCULOSKELETAL DISORDERS                                   
#> 57             MUSCULOSKELETAL DISORDERS                                   
#> 58                                                                         
#> 59                NEUROLOGICAL DISORDERS     10(2.22%)       10/321 = 3.12%
#> 60                NEUROLOGICAL DISORDERS                                   
#> 61                NEUROLOGICAL DISORDERS                                   
#> 62                NEUROLOGICAL DISORDERS                                   
#> 63                NEUROLOGICAL DISORDERS                                   
#> 64                NEUROLOGICAL DISORDERS                                   
#> 65                                                                         
#> 66           RENAL AND URINARY DISORDERS   165(36.59%)     105/321 = 32.71%
#> 67           RENAL AND URINARY DISORDERS                                   
#> 68           RENAL AND URINARY DISORDERS                                   
#> 69           RENAL AND URINARY DISORDERS                                   
#> 70           RENAL AND URINARY DISORDERS                                   
#> 71           RENAL AND URINARY DISORDERS                                   
#> 72           RENAL AND URINARY DISORDERS                                   
#> 73                                                                         
#> 74         REPRODUCTIVE SYSTEM DISORDERS      8(1.77%)        7/321 = 2.18%
#> 75         REPRODUCTIVE SYSTEM DISORDERS                                   
#> 76         REPRODUCTIVE SYSTEM DISORDERS                                   
#> 77         REPRODUCTIVE SYSTEM DISORDERS                                   
#> 78         REPRODUCTIVE SYSTEM DISORDERS                                   
#> 79                                                                         
#> 80           RESPIRATORY TRACT DISORDERS     35(7.76%)      33/321 = 10.28%
#> 81           RESPIRATORY TRACT DISORDERS                                   
#> 82           RESPIRATORY TRACT DISORDERS                                   
#> 83           RESPIRATORY TRACT DISORDERS                                   
#> 84           RESPIRATORY TRACT DISORDERS                                   
#> 85           RESPIRATORY TRACT DISORDERS                                   
#> 86           RESPIRATORY TRACT DISORDERS                                   
#> 87           RESPIRATORY TRACT DISORDERS                                   
#> 88           RESPIRATORY TRACT DISORDERS                                   
#> 89           RESPIRATORY TRACT DISORDERS                                   
#> 90           RESPIRATORY TRACT DISORDERS                                   
#> 91           RESPIRATORY TRACT DISORDERS                                   
#> 92                                                                         
#> 93         SKIN AND APPENDAGES DISORDERS      7(1.55%)        7/321 = 2.18%
#> 94         SKIN AND APPENDAGES DISORDERS                                   
#> 95         SKIN AND APPENDAGES DISORDERS                                   
#> 96         SKIN AND APPENDAGES DISORDERS                                   
#> 97                                                                         
#> 98                    SYSTEMIC DISORDERS    77(17.07%)      56/321 = 17.45%
#> 99                    SYSTEMIC DISORDERS                                   
#> 100                   SYSTEMIC DISORDERS                                   
#> 101                   SYSTEMIC DISORDERS                                   
#> 102                   SYSTEMIC DISORDERS                                   
#> 103                   SYSTEMIC DISORDERS                                   
#> 104                   SYSTEMIC DISORDERS                                   
#> 105                   SYSTEMIC DISORDERS                                   
#> 106                   SYSTEMIC DISORDERS                                   
#> 107                                                                        
#> 108                        UNCODED SIGNS      2(0.44%)        2/321 = 0.62%
#> 109                                                                        
#> 110                                  ALL     235(100%)     142/321 = 44.24%
#> 111                                                                        
#> 112                BEHAVIOURAL DISORDERS      2(0.85%)        1/321 = 0.31%
#> 113                                                                        
#> 114 BLOOD AND LYMPHATIC SYSTEM DISORDERS      8(3.40%)        7/321 = 2.18%
#> 115 BLOOD AND LYMPHATIC SYSTEM DISORDERS                                   
#> 116 BLOOD AND LYMPHATIC SYSTEM DISORDERS                                   
#> 117 BLOOD AND LYMPHATIC SYSTEM DISORDERS                                   
#> 118                                                                        
#> 119     CARDIO-VASCULAR SYSTEM DISORDERS      9(3.83%)        7/321 = 2.18%
#> 120     CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 121     CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 122     CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 123                                                                        
#> 124            DIGESTIVE TRACT DISORDERS    37(15.74%)      34/321 = 10.59%
#> 125            DIGESTIVE TRACT DISORDERS                                   
#> 126            DIGESTIVE TRACT DISORDERS                                   
#> 127            DIGESTIVE TRACT DISORDERS                                   
#> 128            DIGESTIVE TRACT DISORDERS                                   
#> 129            DIGESTIVE TRACT DISORDERS                                   
#> 130            DIGESTIVE TRACT DISORDERS                                   
#> 131            DIGESTIVE TRACT DISORDERS                                   
#> 132            DIGESTIVE TRACT DISORDERS                                   
#> 133            DIGESTIVE TRACT DISORDERS                                   
#> 134            DIGESTIVE TRACT DISORDERS                                   
#> 135            DIGESTIVE TRACT DISORDERS                                   
#> 136            DIGESTIVE TRACT DISORDERS                                   
#> 137                                                                        
#> 138          EAR AND LABYRINTH DISORDERS      2(0.85%)        2/321 = 0.62%
#> 139                                                                        
#> 140                        EYE DISORDERS      3(1.28%)        3/321 = 0.93%
#> 141                                                                        
#> 142             HEPATO-BILIARY DISORDERS      2(0.85%)        2/321 = 0.62%
#> 143                                                                        
#> 144              MAMMARY GLAND DISORDERS      1(0.43%)        1/321 = 0.31%
#> 145                                                                        
#> 146   METABOLISM AND NUTRITION DISORDERS      3(1.28%)        2/321 = 0.62%
#> 147   METABOLISM AND NUTRITION DISORDERS                                   
#> 148                                                                        
#> 149            MUSCULOSKELETAL DISORDERS      7(2.98%)        6/321 = 1.87%
#> 150            MUSCULOSKELETAL DISORDERS                                   
#> 151            MUSCULOSKELETAL DISORDERS                                   
#> 152            MUSCULOSKELETAL DISORDERS                                   
#> 153            MUSCULOSKELETAL DISORDERS                                   
#> 154            MUSCULOSKELETAL DISORDERS                                   
#> 155                                                                        
#> 156               NEUROLOGICAL DISORDERS      5(2.13%)        5/321 = 1.56%
#> 157               NEUROLOGICAL DISORDERS                                   
#> 158               NEUROLOGICAL DISORDERS                                   
#> 159                                                                        
#> 160          RENAL AND URINARY DISORDERS    81(34.47%)      64/321 = 19.94%
#> 161          RENAL AND URINARY DISORDERS                                   
#> 162          RENAL AND URINARY DISORDERS                                   
#> 163          RENAL AND URINARY DISORDERS                                   
#> 164                                                                        
#> 165        REPRODUCTIVE SYSTEM DISORDERS      2(0.85%)        2/321 = 0.62%
#> 166        REPRODUCTIVE SYSTEM DISORDERS                                   
#> 167                                                                        
#> 168          RESPIRATORY TRACT DISORDERS     23(9.79%)       23/321 = 7.17%
#> 169          RESPIRATORY TRACT DISORDERS                                   
#> 170          RESPIRATORY TRACT DISORDERS                                   
#> 171          RESPIRATORY TRACT DISORDERS                                   
#> 172          RESPIRATORY TRACT DISORDERS                                   
#> 173          RESPIRATORY TRACT DISORDERS                                   
#> 174          RESPIRATORY TRACT DISORDERS                                   
#> 175          RESPIRATORY TRACT DISORDERS                                   
#> 176          RESPIRATORY TRACT DISORDERS                                   
#> 177          RESPIRATORY TRACT DISORDERS                                   
#> 178                                                                        
#> 179        SKIN AND APPENDAGES DISORDERS      5(2.13%)        5/321 = 1.56%
#> 180        SKIN AND APPENDAGES DISORDERS                                   
#> 181        SKIN AND APPENDAGES DISORDERS                                   
#> 182        SKIN AND APPENDAGES DISORDERS                                   
#> 183                                                                        
#> 184                   SYSTEMIC DISORDERS    44(18.72%)      38/321 = 11.84%
#> 185                   SYSTEMIC DISORDERS                                   
#> 186                   SYSTEMIC DISORDERS                                   
#> 187                   SYSTEMIC DISORDERS                                   
#> 188                   SYSTEMIC DISORDERS                                   
#> 189                   SYSTEMIC DISORDERS                                   
#> 190                   SYSTEMIC DISORDERS                                   
#> 191                   SYSTEMIC DISORDERS                                   
#> 192                                                                        
#> 193                        UNCODED SIGNS      1(0.43%)        1/321 = 0.31%
#> 194                                                                        
#> 195                                  ALL     216(100%)      139/321 = 43.3%
#> 196                                                                        
#> 197                BEHAVIOURAL DISORDERS      1(0.46%)        1/321 = 0.31%
#> 198                                                                        
#> 199 BLOOD AND LYMPHATIC SYSTEM DISORDERS      2(0.93%)        1/321 = 0.31%
#> 200                                                                        
#> 201     CARDIO-VASCULAR SYSTEM DISORDERS     12(5.56%)       12/321 = 3.74%
#> 202     CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 203     CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 204     CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 205     CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 206     CARDIO-VASCULAR SYSTEM DISORDERS                                   
#> 207                                                                        
#> 208            DIGESTIVE TRACT DISORDERS    35(16.20%)       31/321 = 9.66%
#> 209            DIGESTIVE TRACT DISORDERS                                   
#> 210            DIGESTIVE TRACT DISORDERS                                   
#> 211            DIGESTIVE TRACT DISORDERS                                   
#> 212            DIGESTIVE TRACT DISORDERS                                   
#> 213            DIGESTIVE TRACT DISORDERS                                   
#> 214            DIGESTIVE TRACT DISORDERS                                   
#> 215            DIGESTIVE TRACT DISORDERS                                   
#> 216            DIGESTIVE TRACT DISORDERS                                   
#> 217                                                                        
#> 218          EAR AND LABYRINTH DISORDERS      1(0.46%)        1/321 = 0.31%
#> 219                                                                        
#> 220                        EYE DISORDERS      7(3.24%)        7/321 = 2.18%
#> 221                        EYE DISORDERS                                   
#> 222                        EYE DISORDERS                                   
#> 223                        EYE DISORDERS                                   
#> 224                                                                        
#> 225             HEPATO-BILIARY DISORDERS      3(1.39%)        3/321 = 0.93%
#> 226                                                                        
#> 227   METABOLISM AND NUTRITION DISORDERS      4(1.85%)        4/321 = 1.25%
#> 228   METABOLISM AND NUTRITION DISORDERS                                   
#> 229                                                                        
#> 230            MUSCULOSKELETAL DISORDERS      8(3.70%)        8/321 = 2.49%
#> 231            MUSCULOSKELETAL DISORDERS                                   
#> 232            MUSCULOSKELETAL DISORDERS                                   
#> 233            MUSCULOSKELETAL DISORDERS                                   
#> 234                                                                        
#> 235               NEUROLOGICAL DISORDERS      5(2.31%)        5/321 = 1.56%
#> 236               NEUROLOGICAL DISORDERS                                   
#> 237               NEUROLOGICAL DISORDERS                                   
#> 238               NEUROLOGICAL DISORDERS                                   
#> 239               NEUROLOGICAL DISORDERS                                   
#> 240                                                                        
#> 241          RENAL AND URINARY DISORDERS    84(38.89%)      67/321 = 20.87%
#> 242          RENAL AND URINARY DISORDERS                                   
#> 243          RENAL AND URINARY DISORDERS                                   
#> 244          RENAL AND URINARY DISORDERS                                   
#> 245          RENAL AND URINARY DISORDERS                                   
#> 246          RENAL AND URINARY DISORDERS                                   
#> 247          RENAL AND URINARY DISORDERS                                   
#> 248                                                                        
#> 249        REPRODUCTIVE SYSTEM DISORDERS      6(2.78%)        5/321 = 1.56%
#> 250        REPRODUCTIVE SYSTEM DISORDERS                                   
#> 251        REPRODUCTIVE SYSTEM DISORDERS                                   
#> 252        REPRODUCTIVE SYSTEM DISORDERS                                   
#> 253                                                                        
#> 254          RESPIRATORY TRACT DISORDERS     12(5.56%)       10/321 = 3.12%
#> 255          RESPIRATORY TRACT DISORDERS                                   
#> 256          RESPIRATORY TRACT DISORDERS                                   
#> 257          RESPIRATORY TRACT DISORDERS                                   
#> 258          RESPIRATORY TRACT DISORDERS                                   
#> 259          RESPIRATORY TRACT DISORDERS                                   
#> 260          RESPIRATORY TRACT DISORDERS                                   
#> 261          RESPIRATORY TRACT DISORDERS                                   
#> 262          RESPIRATORY TRACT DISORDERS                                   
#> 263                                                                        
#> 264        SKIN AND APPENDAGES DISORDERS      2(0.93%)        2/321 = 0.62%
#> 265        SKIN AND APPENDAGES DISORDERS                                   
#> 266                                                                        
#> 267                   SYSTEMIC DISORDERS    33(15.28%)       29/321 = 9.03%
#> 268                   SYSTEMIC DISORDERS                                   
#> 269                   SYSTEMIC DISORDERS                                   
#> 270                   SYSTEMIC DISORDERS                                   
#> 271                   SYSTEMIC DISORDERS                                   
#> 272                   SYSTEMIC DISORDERS                                   
#> 273                   SYSTEMIC DISORDERS                                   
#> 274                   SYSTEMIC DISORDERS                                   
#> 275                                                                        
#> 276                        UNCODED SIGNS      1(0.46%)        1/321 = 0.31%
#>                              Prefered Terms n (%) PTNAME At least one PTNAME
#> 1                                       ALL    451(100%)    195/321 = 60.75%
#> 2                                                                           
#> 3                                   ANXIETY     3(0.67%)       1/321 = 0.31%
#> 4                                                                           
#> 5                               ANAEMIA NOS     1(0.22%)       1/321 = 0.31%
#> 6                             LYMPHADENITIS     1(0.22%)       1/321 = 0.31%
#> 7               RED BLOOD CELL DISORDER NOS     2(0.44%)       1/321 = 0.31%
#> 8                        HAEMOLYTIC ANAEMIA     1(0.22%)       1/321 = 0.31%
#> 9                              LEUCOCYTOSIS     5(1.11%)       5/321 = 1.56%
#> 10                                                                          
#> 11                HEART DISORDER AGGRAVATED     2(0.44%)       2/321 = 0.62%
#> 12                              HEART BLOCK     2(0.44%)       2/321 = 0.62%
#> 13                              HYPOTENSION     5(1.11%)       5/321 = 1.56%
#> 14                             HYPERTENSION     3(0.67%)       2/321 = 0.62%
#> 15                           CARDIAC ARREST     2(0.44%)       2/321 = 0.62%
#> 16                     CARDIAC DISORDER NOS     1(0.22%)       1/321 = 0.31%
#> 17                               ARRHYTHMIA     5(1.11%)       4/321 = 1.25%
#> 18                        VALVULAR DISORDER     1(0.22%)       1/321 = 0.31%
#> 19                                                                          
#> 20                                  COLITIS     1(0.22%)       1/321 = 0.31%
#> 21                        GINGIVAL DISORDER     2(0.44%)       1/321 = 0.31%
#> 22             DIGESTIVE TRACT NEOPLASM NOS     1(0.22%)       1/321 = 0.31%
#> 23                                   EMESIS    27(5.99%)      23/321 = 7.17%
#> 24                                ENTERITIS     1(0.22%)       1/321 = 0.31%
#> 25                        ANAL SAC DISORDER     1(0.22%)       1/321 = 0.31%
#> 26                                   NAUSEA     1(0.22%)       1/321 = 0.31%
#> 27             HAEMORRHAGIC GASTROENTERITIS     1(0.22%)       1/321 = 0.31%
#> 28                          GASTROENTERITIS     5(1.11%)       5/321 = 1.56%
#> 29                        INTESTINAL STASIS     4(0.89%)       3/321 = 0.93%
#> 30                   HAEMORRHAGIC DIARRHOEA     2(0.44%)       2/321 = 0.62%
#> 31                                DIARRHOEA    17(3.77%)      15/321 = 4.67%
#> 32             DIGESTIVE TRACT DISORDER NOS     5(1.11%)       5/321 = 1.56%
#> 33                           TOOTH DISORDER     2(0.44%)       2/321 = 0.62%
#> 34                               STOMATITIS     2(0.44%)       2/321 = 0.62%
#> 35                                                                          
#> 36                               OTITIS NOS     2(0.44%)       2/321 = 0.62%
#> 37                           OTITIS EXTERNA     1(0.22%)       1/321 = 0.31%
#> 38                                                                          
#> 39                                 GLAUCOMA     1(0.22%)       1/321 = 0.31%
#> 40                            CORNEAL ULCER     1(0.22%)       1/321 = 0.31%
#> 41                                KERATITIS     1(0.22%)       1/321 = 0.31%
#> 42                           CONJUNCTIVITIS     7(1.55%)       7/321 = 2.18%
#> 43                                                                          
#> 44                              HEPATOPATHY     5(1.11%)       4/321 = 1.25%
#> 45                                                                          
#> 46               MAMMARY GLAND NEOPLASM NOS     1(0.22%)       1/321 = 0.31%
#> 47                                                                          
#> 48                       HYPERPHOSPHATAEMIA     2(0.44%)       2/321 = 0.62%
#> 49                     ELECTROLYTE DISORDER     5(1.11%)       4/321 = 1.25%
#> 50                                                                          
#> 51                                ARTHROSIS     2(0.44%)       2/321 = 0.62%
#> 52             MUSCULOSKELETAL DISORDER NOS     4(0.89%)       3/321 = 0.93%
#> 53                                ARTHRITIS     1(0.22%)       1/321 = 0.31%
#> 54                                 LAMENESS     4(0.89%)       4/321 = 1.25%
#> 55              BONE AND JOINT DISORDER NOS     1(0.22%)       1/321 = 0.31%
#> 56                      MUSCLE WEAKNESS NOS     1(0.22%)       1/321 = 0.31%
#> 57                             OSTEOSARCOMA     2(0.44%)       1/321 = 0.31%
#> 58                                                                          
#> 59                               CONVULSION     2(0.44%)       2/321 = 0.62%
#> 60                   CRANIAL NERVE DISORDER     1(0.22%)       1/321 = 0.31%
#> 61                            MUSCLE TREMOR     1(0.22%)       1/321 = 0.31%
#> 62                        EPILEPTIC SEIZURE     1(0.22%)       1/321 = 0.31%
#> 63                                MYDRIASIS     1(0.22%)       1/321 = 0.31%
#> 64                                   ATAXIA     4(0.89%)       4/321 = 1.25%
#> 65                                                                          
#> 66                      URINE ABNORMALITIES     3(0.67%)       2/321 = 0.62%
#> 67                                 CYSTITIS     5(1.11%)       5/321 = 1.56%
#> 68                      RENAL INSUFFICIENCY  144(31.93%)     96/321 = 29.91%
#> 69               URINARY TRACT DISORDER NOS     2(0.44%)       2/321 = 0.62%
#> 70                                 POLYURIA     4(0.89%)       4/321 = 1.25%
#> 71                     URINARY INCONTINENCE     6(1.33%)       6/321 = 1.87%
#> 72                             UROLITHIASIS     1(0.22%)       1/321 = 0.31%
#> 73                                                                          
#> 74                          PSEUDOPREGNANCY     1(0.22%)       1/321 = 0.31%
#> 75                                 METRITIS     2(0.44%)       2/321 = 0.62%
#> 76   FEMALE REPRODUCTIVE TRACT NEOPLASM NOS     2(0.44%)       1/321 = 0.31%
#> 77                   PROSTATIC DISORDER NOS     2(0.44%)       2/321 = 0.62%
#> 78     MALE REPRODUCTIVE TRACT NEOPLASM NOS     1(0.22%)       1/321 = 0.31%
#> 79                                                                          
#> 80                               TRACHEITIS     3(0.67%)       3/321 = 0.93%
#> 81                        TRACHEAL COLLAPSE     1(0.22%)       1/321 = 0.31%
#> 82      TRACHEAL AND LARYNGEAL DISORDER NOS     5(1.11%)       5/321 = 1.56%
#> 83                                 SNEEZING     1(0.22%)       1/321 = 0.31%
#> 84          RESPIRATORY TRACT INFECTION NOS     4(0.89%)       4/321 = 1.25%
#> 85                                    COUGH     3(0.67%)       3/321 = 0.93%
#> 86                                   ANOXIA     1(0.22%)       1/321 = 0.31%
#> 87                                 DYSPNOEA     3(0.67%)       3/321 = 0.93%
#> 88                                PNEUMONIA     1(0.22%)       1/321 = 0.31%
#> 89                         PULMONARY OEDEMA     9(2.00%)        9/321 = 2.8%
#> 90                               BRONCHITIS     3(0.67%)       3/321 = 0.93%
#> 91                       PULMONARY NEOPLASM     1(0.22%)       1/321 = 0.31%
#> 92                                                                          
#> 93                                 ERYTHEMA     1(0.22%)       1/321 = 0.31%
#> 94                    DERMATITIS AND ECZEMA     2(0.44%)       2/321 = 0.62%
#> 95                     SKIN TEXTURAL CHANGE     1(0.22%)       1/321 = 0.31%
#> 96                 BACTERIAL SKIN INFECTION     3(0.67%)       3/321 = 0.93%
#> 97                                                                          
#> 98                             HYPERTHERMIA     2(0.44%)       2/321 = 0.62%
#> 99                              WEIGHT LOSS     5(1.11%)       5/321 = 1.56%
#> 100                                   DEATH    14(3.10%)      14/321 = 4.36%
#> 101                                ANOREXIA    26(5.76%)      25/321 = 7.79%
#> 102 SUSPECTED INFECTIOUS AGENT TRANSMISSION     1(0.22%)       1/321 = 0.31%
#> 103                   APPETITE DISORDER NOS     1(0.22%)       1/321 = 0.31%
#> 104                             DEHYDRATION     5(1.11%)       5/321 = 1.56%
#> 105                              POLYDIPSIA    10(2.22%)      10/321 = 3.12%
#> 106                                LETHARGY    13(2.88%)      12/321 = 3.74%
#> 107                                                                         
#> 108                            UNCODED SIGN     2(0.44%)       2/321 = 0.62%
#> 109                                                                         
#> 110                                     ALL    235(100%)    142/321 = 44.24%
#> 111                                                                         
#> 112                                 ANXIETY     2(0.85%)       1/321 = 0.31%
#> 113                                                                         
#> 114                             ANAEMIA NOS     1(0.43%)       1/321 = 0.31%
#> 115                           LYMPHADENITIS     1(0.43%)       1/321 = 0.31%
#> 116                      HAEMOLYTIC ANAEMIA     1(0.43%)       1/321 = 0.31%
#> 117                            LEUCOCYTOSIS     5(2.13%)       5/321 = 1.56%
#> 118                                                                         
#> 119                             HYPOTENSION     2(0.85%)       2/321 = 0.62%
#> 120                          CARDIAC ARREST     2(0.85%)       2/321 = 0.62%
#> 121                            HYPERTENSION     3(1.28%)       2/321 = 0.62%
#> 122                              ARRHYTHMIA     2(0.85%)       2/321 = 0.62%
#> 123                                                                         
#> 124                         GASTROENTERITIS     3(1.28%)       3/321 = 0.93%
#> 125                       GINGIVAL DISORDER     1(0.43%)       1/321 = 0.31%
#> 126                                  NAUSEA     1(0.43%)       1/321 = 0.31%
#> 127                               DIARRHOEA     8(3.40%)       8/321 = 2.49%
#> 128            DIGESTIVE TRACT DISORDER NOS     2(0.85%)       2/321 = 0.62%
#> 129                              STOMATITIS     2(0.85%)       2/321 = 0.62%
#> 130                       INTESTINAL STASIS     1(0.43%)       1/321 = 0.31%
#> 131                                 COLITIS     1(0.43%)       1/321 = 0.31%
#> 132            DIGESTIVE TRACT NEOPLASM NOS     1(0.43%)       1/321 = 0.31%
#> 133                       ANAL SAC DISORDER     1(0.43%)       1/321 = 0.31%
#> 134                          TOOTH DISORDER     1(0.43%)       1/321 = 0.31%
#> 135                                  EMESIS    14(5.96%)      13/321 = 4.05%
#> 136                               ENTERITIS     1(0.43%)       1/321 = 0.31%
#> 137                                                                         
#> 138                              OTITIS NOS     2(0.85%)       2/321 = 0.62%
#> 139                                                                         
#> 140                          CONJUNCTIVITIS     3(1.28%)       3/321 = 0.93%
#> 141                                                                         
#> 142                             HEPATOPATHY     2(0.85%)       2/321 = 0.62%
#> 143                                                                         
#> 144              MAMMARY GLAND NEOPLASM NOS     1(0.43%)       1/321 = 0.31%
#> 145                                                                         
#> 146                    ELECTROLYTE DISORDER     2(0.85%)       1/321 = 0.31%
#> 147                      HYPERPHOSPHATAEMIA     1(0.43%)       1/321 = 0.31%
#> 148                                                                         
#> 149                                LAMENESS     1(0.43%)       1/321 = 0.31%
#> 150                               ARTHROSIS     2(0.85%)       2/321 = 0.62%
#> 151             BONE AND JOINT DISORDER NOS     1(0.43%)       1/321 = 0.31%
#> 152                     MUSCLE WEAKNESS NOS     1(0.43%)       1/321 = 0.31%
#> 153            MUSCULOSKELETAL DISORDER NOS     1(0.43%)       1/321 = 0.31%
#> 154                            OSTEOSARCOMA     1(0.43%)       1/321 = 0.31%
#> 155                                                                         
#> 156                                  ATAXIA     3(1.28%)       3/321 = 0.93%
#> 157                              CONVULSION     1(0.43%)       1/321 = 0.31%
#> 158                           MUSCLE TREMOR     1(0.43%)       1/321 = 0.31%
#> 159                                                                         
#> 160                                POLYURIA     2(0.85%)       2/321 = 0.62%
#> 161                     URINE ABNORMALITIES     1(0.43%)       1/321 = 0.31%
#> 162                     RENAL INSUFFICIENCY   75(31.91%)     60/321 = 18.69%
#> 163                                CYSTITIS     3(1.28%)       3/321 = 0.93%
#> 164                                                                         
#> 165                  PROSTATIC DISORDER NOS     1(0.43%)       1/321 = 0.31%
#> 166                         PSEUDOPREGNANCY     1(0.43%)       1/321 = 0.31%
#> 167                                                                         
#> 168                                  ANOXIA     1(0.43%)       1/321 = 0.31%
#> 169                                   COUGH     2(0.85%)       2/321 = 0.62%
#> 170                        PULMONARY OEDEMA     6(2.55%)       6/321 = 1.87%
#> 171                              BRONCHITIS     2(0.85%)       2/321 = 0.62%
#> 172         RESPIRATORY TRACT INFECTION NOS     3(1.28%)       3/321 = 0.93%
#> 173                               PNEUMONIA     1(0.43%)       1/321 = 0.31%
#> 174                       TRACHEAL COLLAPSE     1(0.43%)       1/321 = 0.31%
#> 175                                DYSPNOEA     2(0.85%)       2/321 = 0.62%
#> 176     TRACHEAL AND LARYNGEAL DISORDER NOS     3(1.28%)       3/321 = 0.93%
#> 177                              TRACHEITIS     2(0.85%)       2/321 = 0.62%
#> 178                                                                         
#> 179                   DERMATITIS AND ECZEMA     1(0.43%)       1/321 = 0.31%
#> 180                                ERYTHEMA     1(0.43%)       1/321 = 0.31%
#> 181                BACTERIAL SKIN INFECTION     2(0.85%)       2/321 = 0.62%
#> 182                    SKIN TEXTURAL CHANGE     1(0.43%)       1/321 = 0.31%
#> 183                                                                         
#> 184                                ANOREXIA    12(5.11%)      12/321 = 3.74%
#> 185 SUSPECTED INFECTIOUS AGENT TRANSMISSION     1(0.43%)       1/321 = 0.31%
#> 186                              POLYDIPSIA     6(2.55%)       6/321 = 1.87%
#> 187                                   DEATH     8(3.40%)       8/321 = 2.49%
#> 188                             DEHYDRATION     2(0.85%)       2/321 = 0.62%
#> 189                             WEIGHT LOSS     4(1.70%)       4/321 = 1.25%
#> 190                            HYPERTHERMIA     1(0.43%)       1/321 = 0.31%
#> 191                                LETHARGY    10(4.26%)      10/321 = 3.12%
#> 192                                                                         
#> 193                            UNCODED SIGN     1(0.43%)       1/321 = 0.31%
#> 194                                                                         
#> 195                                     ALL    216(100%)     139/321 = 43.3%
#> 196                                                                         
#> 197                                 ANXIETY     1(0.46%)       1/321 = 0.31%
#> 198                                                                         
#> 199             RED BLOOD CELL DISORDER NOS     2(0.93%)       1/321 = 0.31%
#> 200                                                                         
#> 201               HEART DISORDER AGGRAVATED     2(0.93%)       2/321 = 0.62%
#> 202                             HEART BLOCK     2(0.93%)       2/321 = 0.62%
#> 203                             HYPOTENSION     3(1.39%)       3/321 = 0.93%
#> 204                       VALVULAR DISORDER     1(0.46%)       1/321 = 0.31%
#> 205                              ARRHYTHMIA     3(1.39%)       3/321 = 0.93%
#> 206                    CARDIAC DISORDER NOS     1(0.46%)       1/321 = 0.31%
#> 207                                                                         
#> 208                  HAEMORRHAGIC DIARRHOEA     2(0.93%)       2/321 = 0.62%
#> 209            DIGESTIVE TRACT DISORDER NOS     3(1.39%)       3/321 = 0.93%
#> 210                                  EMESIS    13(6.02%)      13/321 = 4.05%
#> 211                               DIARRHOEA     9(4.17%)       8/321 = 2.49%
#> 212            HAEMORRHAGIC GASTROENTERITIS     1(0.46%)       1/321 = 0.31%
#> 213                         GASTROENTERITIS     2(0.93%)       2/321 = 0.62%
#> 214                       GINGIVAL DISORDER     1(0.46%)       1/321 = 0.31%
#> 215                       INTESTINAL STASIS     3(1.39%)       3/321 = 0.93%
#> 216                          TOOTH DISORDER     1(0.46%)       1/321 = 0.31%
#> 217                                                                         
#> 218                          OTITIS EXTERNA     1(0.46%)       1/321 = 0.31%
#> 219                                                                         
#> 220                                GLAUCOMA     1(0.46%)       1/321 = 0.31%
#> 221                          CONJUNCTIVITIS     4(1.85%)       4/321 = 1.25%
#> 222                           CORNEAL ULCER     1(0.46%)       1/321 = 0.31%
#> 223                               KERATITIS     1(0.46%)       1/321 = 0.31%
#> 224                                                                         
#> 225                             HEPATOPATHY     3(1.39%)       3/321 = 0.93%
#> 226                                                                         
#> 227                    ELECTROLYTE DISORDER     3(1.39%)       3/321 = 0.93%
#> 228                      HYPERPHOSPHATAEMIA     1(0.46%)       1/321 = 0.31%
#> 229                                                                         
#> 230                            OSTEOSARCOMA     1(0.46%)       1/321 = 0.31%
#> 231                               ARTHRITIS     1(0.46%)       1/321 = 0.31%
#> 232                                LAMENESS     3(1.39%)       3/321 = 0.93%
#> 233            MUSCULOSKELETAL DISORDER NOS     3(1.39%)       3/321 = 0.93%
#> 234                                                                         
#> 235                              CONVULSION     1(0.46%)       1/321 = 0.31%
#> 236                  CRANIAL NERVE DISORDER     1(0.46%)       1/321 = 0.31%
#> 237                               MYDRIASIS     1(0.46%)       1/321 = 0.31%
#> 238                       EPILEPTIC SEIZURE     1(0.46%)       1/321 = 0.31%
#> 239                                  ATAXIA     1(0.46%)       1/321 = 0.31%
#> 240                                                                         
#> 241                     URINE ABNORMALITIES     2(0.93%)       2/321 = 0.62%
#> 242                     RENAL INSUFFICIENCY   69(31.94%)     60/321 = 18.69%
#> 243                            UROLITHIASIS     1(0.46%)       1/321 = 0.31%
#> 244                                POLYURIA     2(0.93%)       2/321 = 0.62%
#> 245              URINARY TRACT DISORDER NOS     2(0.93%)       2/321 = 0.62%
#> 246                                CYSTITIS     2(0.93%)       2/321 = 0.62%
#> 247                    URINARY INCONTINENCE     6(2.78%)       6/321 = 1.87%
#> 248                                                                         
#> 249    MALE REPRODUCTIVE TRACT NEOPLASM NOS     1(0.46%)       1/321 = 0.31%
#> 250                  PROSTATIC DISORDER NOS     1(0.46%)       1/321 = 0.31%
#> 251                                METRITIS     2(0.93%)       2/321 = 0.62%
#> 252  FEMALE REPRODUCTIVE TRACT NEOPLASM NOS     2(0.93%)       1/321 = 0.31%
#> 253                                                                         
#> 254                              BRONCHITIS     1(0.46%)       1/321 = 0.31%
#> 255                        PULMONARY OEDEMA     3(1.39%)       3/321 = 0.93%
#> 256         RESPIRATORY TRACT INFECTION NOS     1(0.46%)       1/321 = 0.31%
#> 257                                SNEEZING     1(0.46%)       1/321 = 0.31%
#> 258                                   COUGH     1(0.46%)       1/321 = 0.31%
#> 259                      PULMONARY NEOPLASM     1(0.46%)       1/321 = 0.31%
#> 260                                DYSPNOEA     1(0.46%)       1/321 = 0.31%
#> 261                              TRACHEITIS     1(0.46%)       1/321 = 0.31%
#> 262     TRACHEAL AND LARYNGEAL DISORDER NOS     2(0.93%)       2/321 = 0.62%
#> 263                                                                         
#> 264                BACTERIAL SKIN INFECTION     1(0.46%)       1/321 = 0.31%
#> 265                   DERMATITIS AND ECZEMA     1(0.46%)       1/321 = 0.31%
#> 266                                                                         
#> 267                                ANOREXIA    14(6.48%)      14/321 = 4.36%
#> 268                                LETHARGY     3(1.39%)       3/321 = 0.93%
#> 269                   APPETITE DISORDER NOS     1(0.46%)       1/321 = 0.31%
#> 270                             DEHYDRATION     3(1.39%)       3/321 = 0.93%
#> 271                            HYPERTHERMIA     1(0.46%)       1/321 = 0.31%
#> 272                              POLYDIPSIA     4(1.85%)       4/321 = 1.25%
#> 273                             WEIGHT LOSS     1(0.46%)       1/321 = 0.31%
#> 274                                   DEATH     6(2.78%)       6/321 = 1.87%
#> 275                                                                         
#> 276                            UNCODED SIGN     1(0.46%)       1/321 = 0.31%
#>     randtrt
#> 1       A+B
#> 2       A+B
#> 3       A+B
#> 4       A+B
#> 5       A+B
#> 6       A+B
#> 7       A+B
#> 8       A+B
#> 9       A+B
#> 10      A+B
#> 11      A+B
#> 12      A+B
#> 13      A+B
#> 14      A+B
#> 15      A+B
#> 16      A+B
#> 17      A+B
#> 18      A+B
#> 19      A+B
#> 20      A+B
#> 21      A+B
#> 22      A+B
#> 23      A+B
#> 24      A+B
#> 25      A+B
#> 26      A+B
#> 27      A+B
#> 28      A+B
#> 29      A+B
#> 30      A+B
#> 31      A+B
#> 32      A+B
#> 33      A+B
#> 34      A+B
#> 35      A+B
#> 36      A+B
#> 37      A+B
#> 38      A+B
#> 39      A+B
#> 40      A+B
#> 41      A+B
#> 42      A+B
#> 43      A+B
#> 44      A+B
#> 45      A+B
#> 46      A+B
#> 47      A+B
#> 48      A+B
#> 49      A+B
#> 50      A+B
#> 51      A+B
#> 52      A+B
#> 53      A+B
#> 54      A+B
#> 55      A+B
#> 56      A+B
#> 57      A+B
#> 58      A+B
#> 59      A+B
#> 60      A+B
#> 61      A+B
#> 62      A+B
#> 63      A+B
#> 64      A+B
#> 65      A+B
#> 66      A+B
#> 67      A+B
#> 68      A+B
#> 69      A+B
#> 70      A+B
#> 71      A+B
#> 72      A+B
#> 73      A+B
#> 74      A+B
#> 75      A+B
#> 76      A+B
#> 77      A+B
#> 78      A+B
#> 79      A+B
#> 80      A+B
#> 81      A+B
#> 82      A+B
#> 83      A+B
#> 84      A+B
#> 85      A+B
#> 86      A+B
#> 87      A+B
#> 88      A+B
#> 89      A+B
#> 90      A+B
#> 91      A+B
#> 92      A+B
#> 93      A+B
#> 94      A+B
#> 95      A+B
#> 96      A+B
#> 97      A+B
#> 98      A+B
#> 99      A+B
#> 100     A+B
#> 101     A+B
#> 102     A+B
#> 103     A+B
#> 104     A+B
#> 105     A+B
#> 106     A+B
#> 107     A+B
#> 108     A+B
#> 109        
#> 110       A
#> 111       A
#> 112       A
#> 113       A
#> 114       A
#> 115       A
#> 116       A
#> 117       A
#> 118       A
#> 119       A
#> 120       A
#> 121       A
#> 122       A
#> 123       A
#> 124       A
#> 125       A
#> 126       A
#> 127       A
#> 128       A
#> 129       A
#> 130       A
#> 131       A
#> 132       A
#> 133       A
#> 134       A
#> 135       A
#> 136       A
#> 137       A
#> 138       A
#> 139       A
#> 140       A
#> 141       A
#> 142       A
#> 143       A
#> 144       A
#> 145       A
#> 146       A
#> 147       A
#> 148       A
#> 149       A
#> 150       A
#> 151       A
#> 152       A
#> 153       A
#> 154       A
#> 155       A
#> 156       A
#> 157       A
#> 158       A
#> 159       A
#> 160       A
#> 161       A
#> 162       A
#> 163       A
#> 164       A
#> 165       A
#> 166       A
#> 167       A
#> 168       A
#> 169       A
#> 170       A
#> 171       A
#> 172       A
#> 173       A
#> 174       A
#> 175       A
#> 176       A
#> 177       A
#> 178       A
#> 179       A
#> 180       A
#> 181       A
#> 182       A
#> 183       A
#> 184       A
#> 185       A
#> 186       A
#> 187       A
#> 188       A
#> 189       A
#> 190       A
#> 191       A
#> 192       A
#> 193       A
#> 194        
#> 195       B
#> 196       B
#> 197       B
#> 198       B
#> 199       B
#> 200       B
#> 201       B
#> 202       B
#> 203       B
#> 204       B
#> 205       B
#> 206       B
#> 207       B
#> 208       B
#> 209       B
#> 210       B
#> 211       B
#> 212       B
#> 213       B
#> 214       B
#> 215       B
#> 216       B
#> 217       B
#> 218       B
#> 219       B
#> 220       B
#> 221       B
#> 222       B
#> 223       B
#> 224       B
#> 225       B
#> 226       B
#> 227       B
#> 228       B
#> 229       B
#> 230       B
#> 231       B
#> 232       B
#> 233       B
#> 234       B
#> 235       B
#> 236       B
#> 237       B
#> 238       B
#> 239       B
#> 240       B
#> 241       B
#> 242       B
#> 243       B
#> 244       B
#> 245       B
#> 246       B
#> 247       B
#> 248       B
#> 249       B
#> 250       B
#> 251       B
#> 252       B
#> 253       B
#> 254       B
#> 255       B
#> 256       B
#> 257       B
#> 258       B
#> 259       B
#> 260       B
#> 261       B
#> 262       B
#> 263       B
#> 264       B
#> 265       B
#> 266       B
#> 267       B
#> 268       B
#> 269       B
#> 270       B
#> 271       B
#> 272       B
#> 273       B
#> 274       B
#> 275       B
#> 276       B
#> 
#> ############################################
#> 


# show formatted results in HTML
ft=report.doc(test,valign=TRUE)
ft

#> a flextable object.
#> col_keys: `System Organ Class`, `n (%) SOCNAME`, `At least one SOCNAME`, `Prefered Terms`, `n (%) PTNAME`, `At least one PTNAME`, `randtrt` 
#> header has 2 row(s) 
#> body has 276 row(s) 
#> original dataset sample: 
#>                     System Organ Class n (%) SOCNAME At least one SOCNAME
#> 1                                  ALL     451(100%)     195/321 = 60.75%
#> 2                                                                        
#> 3                BEHAVIOURAL DISORDERS      3(0.67%)        1/321 = 0.31%
#> 4                                                                        
#> 5 BLOOD AND LYMPHATIC SYSTEM DISORDERS     10(2.22%)        7/321 = 2.18%
#>   Prefered Terms n (%) PTNAME At least one PTNAME randtrt
#> 1            ALL    451(100%)    195/321 = 60.75%     A+B
#> 2                                                     A+B
#> 3        ANXIETY     3(0.67%)       1/321 = 0.31%     A+B
#> 4                                                     A+B
#> 5    ANAEMIA NOS     1(0.22%)       1/321 = 0.31%     A+B