Creates a desc object for "quantitative" statistics reporting

report.quanti Returns quantitative descriptive statistics such as mean, median, standard deviation etc...

report.quanti(data, y, x1 = NULL, x2 = NULL, y.label = y,
  round = 2, total = F, scientific = F, digits = NULL,
  at.row = NULL, subjid = NULL, geomean = F, add.mad = F,
  default.stat = T, func.stat, stat.name = "Statistics",
  func.stat.name = "")

Arguments

data
Data.frame object

y
Character indicating a numerical vector in the data frame passed to data argument

x1
Character indicating a factor in the data (levels will be displayed in columns)

x2
Character indicating a factor in the data (levels will be displayed in lines)

y.label
Character Indicates the label for y parameter

round
Numeric to indicate how to round statistics

total
Logical to indicate if a "Total" column should be added

scientific
Logical Indicates if statistics should be displayed in scientific notations or not

digits
Numeric (used if scientifc=TRUE) to indicate how many digits to use in scientific notation

at.row
Character Used to space the results (see examples)

subjid
Character Indicates the column in which there is the subject Id to add the number of subjects in the column header if x1 and x2 are not null.

geomean
Logical If yes geometric mean is calculated instead of arithmetic mean: exp(mean(log(x),na.rm=T)) fpr x>0

add.mad
Logical If yes the Median Absolute Deviance is added to the median statistics (see function mad)

default.stat
Logical (default to TRUE). If FALSE you can specify your own example

func.stat
Function. Used only if default.stat=FALSE

stat.name
Character. Indicates the name of the variable that report the statistics Default = "Statistics"

func.stat.name
Character. Used only if default.stat=FALSE. Indicates the name of specific statistic you want to report

Value

A desc object.

Details

This function computes and reports quantitative statistics on y. And can gives the statistics by level of two factors (x1 in columns and/or x2 in rows). See the example to show the results. If total=TRUE, the last column is the statistics performed overall levels of x1 for each levels of x2. Quantiles are calculated using type 3 (SAS) algorithms.

"geomean" compute the geometric mean defined as exp(mean(log(y))). The values below or equal 0 are removed and a message is printed to indicate how many values were deleted to calculate the geometric mean.

N returns the number of observations (including NA values)

See also

Examples


data(data)

# Quantitative statistics with no factor

report.quanti(data=data,y="y_numeric",total=TRUE,y.label="Awesome results")

#>   Statistics Awesome results        Total
#> 6          N             396          396
#> 3  Mean (SD)      2.56(2.20)   2.56(2.20)
#> 4     Median            2.71         2.71
#> 2    [Q1;Q3]     [1.04;4.33]  [1.04;4.33]
#> 1  [Min;Max]    [-2.99;7.96] [-2.99;7.96]
#> 5    Missing              10           10


#' # Quantitative statistics with no factor with geometric mean (option geomean=TRUE)

report.quanti(data=data,y="y_numeric",y.label="Awesome results",geomean=TRUE)

#> 66 values were removed to calculate the Geometric mean

#>      Statistics Awesome results
#> 1             N             396
#> 3 Geo Mean (SD)      2.52(2.20)
#> 6        Median            2.71
#> 4       [Q1;Q3]     [1.04;4.33]
#> 5     [Min;Max]    [-2.99;7.96]
#> 2       Missing              10


# Quantitative statistics with one factor

report.quanti(data=data,y="y_numeric",x1="GROUP")

#>   Statistics            A            B            C
#> 1          N          180          120           96
#> 3  Mean (SD)   1.46(1.50)   3.15(2.00)   3.87(2.52)
#> 6     Median         1.59         3.75         4.73
#> 4    [Q1;Q3]  [0.45;2.50]  [2.46;4.44]  [3.44;5.30]
#> 5  [Min;Max] [-2.34;4.36] [-2.44;6.19] [-2.99;7.96]
#> 2    Missing            4            4            2


# One factor with total column

report.quanti(data=data,y="y_numeric",x1="GROUP",total=TRUE)

#>   Statistics            A            B            C        Total
#> 6          N          180          120           96          396
#> 3  Mean (SD)   1.46(1.50)   3.15(2.00)   3.87(2.52)   2.56(2.20)
#> 4     Median         1.59         3.75         4.73         2.71
#> 2    [Q1;Q3]  [0.45;2.50]  [2.46;4.44]  [3.44;5.30]  [1.04;4.33]
#> 1  [Min;Max] [-2.34;4.36] [-2.44;6.19] [-2.99;7.96] [-2.99;7.96]
#> 5    Missing            4            4            2           10


# Quantitative statistics with two factors

report.quanti(data=data,y="y_numeric",x1="GROUP",x2="TIMEPOINT")

#>    TIMEPOINT Statistics             A             B             C
#> 1         D0          N            30            20            16
#> 3         D0  Mean (SD)   -0.93(0.86)   -0.67(1.09)   -1.19(0.92)
#> 6         D0     Median         -0.82         -0.69         -1.26
#> 4         D0    [Q1;Q3] [-1.59;-0.16] [-1.39;-0.06] [-1.62;-0.83]
#> 5         D0  [Min;Max]  [-2.34;0.36]  [-2.44;2.10]  [-2.99;0.66]
#> 2         D0    Missing             1             1             0
#> 7         D1          N            30            20            16
#> 9         D1  Mean (SD)    1.83(1.04)    4.17(1.28)    4.98(0.69)
#> 12        D1     Median          1.78          4.19          5.08
#> 10        D1    [Q1;Q3] [ 0.94; 2.54] [ 3.23; 4.92] [ 4.58; 5.46]
#> 11        D1  [Min;Max]  [ 0.11;3.88]  [ 1.48;6.19]  [ 3.80;6.23]
#> 8         D1    Missing             1             0             0
#> 13        D2          N            30            20            16
#> 15        D2  Mean (SD)    1.97(1.17)    4.04(0.89)    4.90(1.36)
#> 18        D2     Median          1.66          4.19          5.06
#> 16        D2    [Q1;Q3] [ 1.23; 2.86] [ 3.62; 4.36] [ 4.34; 5.20]
#> 17        D2  [Min;Max]  [-0.18;4.36]  [ 2.03;5.63]  [ 2.39;7.96]
#> 14        D2    Missing             1             1             0
#> 19        D3          N            30            20            16
#> 21        D3  Mean (SD)    1.78(1.17)    3.81(0.94)    5.07(1.12)
#> 24        D3     Median          1.78          3.63          5.22
#> 22        D3    [Q1;Q3] [ 0.93; 2.42] [ 3.13; 4.44] [ 4.11; 5.66]
#> 23        D3  [Min;Max]  [-0.16;3.90]  [ 2.46;6.01]  [ 3.16;7.37]
#> 20        D3    Missing             0             1             1
#> 25        D4          N            30            20            16
#> 27        D4  Mean (SD)    1.83(0.85)    3.80(0.95)    5.17(1.03)
#> 30        D4     Median          1.67          3.83          4.88
#> 28        D4    [Q1;Q3] [ 1.26; 2.32] [ 3.12; 4.42] [ 4.69; 5.50]
#> 29        D4  [Min;Max]  [ 0.38;3.97]  [ 2.31;5.41]  [ 3.24;6.96]
#> 26        D4    Missing             1             1             1
#> 31        D5          N            30            20            16
#> 33        D5  Mean (SD)    2.27(1.20)    3.64(1.19)    4.43(0.98)
#> 36        D5     Median          2.50          3.86          4.57
#> 34        D5    [Q1;Q3] [ 1.77; 3.21] [ 2.59; 4.60] [ 3.44; 4.97]
#> 35        D5  [Min;Max]  [-1.19;4.31]  [ 0.91;5.12]  [ 2.95;6.54]
#> 32        D5    Missing             0             0             0


# Quantitative statistics with two factors and a total column

report.quanti(data=data,y="y_numeric",x1="GROUP",x2="TIMEPOINT",total=TRUE)

#>    TIMEPOINT Statistics             A             B             C         Total
#> 6         D0          N            30            20            16            66
#> 3         D0  Mean (SD)   -0.93(0.86)   -0.67(1.09)   -1.19(0.92)   -0.92(0.95)
#> 4         D0     Median         -0.82         -0.69         -1.26         -0.86
#> 2         D0    [Q1;Q3] [-1.59;-0.16] [-1.39;-0.06] [-1.62;-0.83] [-1.55;-0.16]
#> 1         D0  [Min;Max]  [-2.34;0.36]  [-2.44;2.10]  [-2.99;0.66]  [-2.99;2.10]
#> 5         D0    Missing             1             1             0             2
#> 12        D1          N            30            20            16            66
#> 9         D1  Mean (SD)    1.83(1.04)    4.17(1.28)    4.98(0.69)    3.33(1.73)
#> 10        D1     Median          1.78          4.19          5.08          3.57
#> 8         D1    [Q1;Q3] [ 0.94; 2.54] [ 3.23; 4.92] [ 4.58; 5.46] [ 1.78; 4.91]
#> 7         D1  [Min;Max]  [ 0.11;3.88]  [ 1.48;6.19]  [ 3.80;6.23]  [ 0.11;6.23]
#> 11        D1    Missing             1             0             0             1
#> 18        D2          N            30            20            16            66
#> 15        D2  Mean (SD)    1.97(1.17)    4.04(0.89)    4.90(1.36)    3.32(1.70)
#> 16        D2     Median          1.66          4.19          5.06          3.57
#> 14        D2    [Q1;Q3] [ 1.23; 2.86] [ 3.62; 4.36] [ 4.34; 5.20] [ 1.89; 4.44]
#> 13        D2  [Min;Max]  [-0.18;4.36]  [ 2.03;5.63]  [ 2.39;7.96]  [-0.18;7.96]
#> 17        D2    Missing             1             1             0             2
#> 24        D3          N            30            20            16            66
#> 21        D3  Mean (SD)    1.78(1.17)    3.81(0.94)    5.07(1.12)    3.15(1.75)
#> 22        D3     Median          1.78          3.63          5.22          3.15
#> 20        D3    [Q1;Q3] [ 0.93; 2.42] [ 3.13; 4.44] [ 4.11; 5.66] [ 1.80; 4.39]
#> 19        D3  [Min;Max]  [-0.16;3.90]  [ 2.46;6.01]  [ 3.16;7.37]  [-0.16;7.37]
#> 23        D3    Missing             0             1             1             2
#> 30        D4          N            30            20            16            66
#> 27        D4  Mean (SD)    1.83(0.85)    3.80(0.95)    5.17(1.03)    3.22(1.66)
#> 28        D4     Median          1.67          3.83          4.88          3.16
#> 26        D4    [Q1;Q3] [ 1.26; 2.32] [ 3.12; 4.42] [ 4.69; 5.50] [ 1.69; 4.48]
#> 25        D4  [Min;Max]  [ 0.38;3.97]  [ 2.31;5.41]  [ 3.24;6.96]  [ 0.38;6.96]
#> 29        D4    Missing             1             1             1             3
#> 36        D5          N            30            20            16            66
#> 33        D5  Mean (SD)    2.27(1.20)    3.64(1.19)    4.43(0.98)    3.21(1.45)
#> 34        D5     Median          2.50          3.86          4.57          3.28
#> 32        D5    [Q1;Q3] [ 1.77; 3.21] [ 2.59; 4.60] [ 3.44; 4.97] [ 2.42; 4.44]
#> 31        D5  [Min;Max]  [-1.19;4.31]  [ 0.91;5.12]  [ 2.95;6.54]  [-1.19;6.54]
#> 35        D5    Missing             0             0             0             0


# Add median absolute deviance to the median statistics

report.quanti(data=data,y="y_numeric",x1="GROUP",x2="TIMEPOINT",total=TRUE,add.mad=TRUE)

#>    TIMEPOINT   Statistics             A             B             C
#> 6         D0            N            30            20            16
#> 3         D0    Mean (SD)   -0.93(0.86)   -0.67(1.09)   -1.19(0.92)
#> 4         D0 Median (MAD)   -0.82(1.00)   -0.69(0.97)   -1.26(0.59)
#> 2         D0      [Q1;Q3] [-1.59;-0.16] [-1.39;-0.06] [-1.62;-0.83]
#> 1         D0    [Min;Max]  [-2.34;0.36]  [-2.44;2.10]  [-2.99;0.66]
#> 5         D0      Missing             1             1             0
#> 12        D1            N            30            20            16
#> 9         D1    Mean (SD)    1.83(1.04)    4.17(1.28)    4.98(0.69)
#> 10        D1 Median (MAD)    1.78(1.12)    4.19(1.25)    5.08(0.66)
#> 8         D1      [Q1;Q3] [ 0.94; 2.54] [ 3.23; 4.92] [ 4.58; 5.46]
#> 7         D1    [Min;Max]  [ 0.11;3.88]  [ 1.48;6.19]  [ 3.80;6.23]
#> 11        D1      Missing             1             0             0
#> 18        D2            N            30            20            16
#> 15        D2    Mean (SD)    1.97(1.17)    4.04(0.89)    4.90(1.36)
#> 16        D2 Median (MAD)    1.66(1.08)    4.19(0.50)    5.06(0.67)
#> 14        D2      [Q1;Q3] [ 1.23; 2.86] [ 3.62; 4.36] [ 4.34; 5.20]
#> 13        D2    [Min;Max]  [-0.18;4.36]  [ 2.03;5.63]  [ 2.39;7.96]
#> 17        D2      Missing             1             1             0
#> 24        D3            N            30            20            16
#> 21        D3    Mean (SD)    1.78(1.17)    3.81(0.94)    5.07(1.12)
#> 22        D3 Median (MAD)    1.78(1.23)    3.63(1.20)    5.22(1.25)
#> 20        D3      [Q1;Q3] [ 0.93; 2.42] [ 3.13; 4.44] [ 4.11; 5.66]
#> 19        D3    [Min;Max]  [-0.16;3.90]  [ 2.46;6.01]  [ 3.16;7.37]
#> 23        D3      Missing             0             1             1
#> 30        D4            N            30            20            16
#> 27        D4    Mean (SD)    1.83(0.85)    3.80(0.95)    5.17(1.03)
#> 28        D4 Median (MAD)    1.67(0.62)    3.83(0.96)    4.88(0.82)
#> 26        D4      [Q1;Q3] [ 1.26; 2.32] [ 3.12; 4.42] [ 4.69; 5.50]
#> 25        D4    [Min;Max]  [ 0.38;3.97]  [ 2.31;5.41]  [ 3.24;6.96]
#> 29        D4      Missing             1             1             1
#> 36        D5            N            30            20            16
#> 33        D5    Mean (SD)    2.27(1.20)    3.64(1.19)    4.43(0.98)
#> 34        D5 Median (MAD)    2.50(1.09)    3.86(1.18)    4.57(0.95)
#> 32        D5      [Q1;Q3] [ 1.77; 3.21] [ 2.59; 4.60] [ 3.44; 4.97]
#> 31        D5    [Min;Max]  [-1.19;4.31]  [ 0.91;5.12]  [ 2.95;6.54]
#> 35        D5      Missing             0             0             0
#>            Total
#> 6             66
#> 3    -0.92(0.95)
#> 4    -0.86(1.03)
#> 2  [-1.55;-0.16]
#> 1   [-2.99;2.10]
#> 5              2
#> 12            66
#> 9     3.33(1.73)
#> 10    3.57(2.14)
#> 8  [ 1.78; 4.91]
#> 7   [ 0.11;6.23]
#> 11             1
#> 18            66
#> 15    3.32(1.70)
#> 16    3.57(1.89)
#> 14 [ 1.89; 4.44]
#> 13  [-0.18;7.96]
#> 17             2
#> 24            66
#> 21    3.15(1.75)
#> 22    3.15(1.88)
#> 20 [ 1.80; 4.39]
#> 19  [-0.16;7.37]
#> 23             2
#> 30            66
#> 27    3.22(1.66)
#> 28    3.16(2.14)
#> 26 [ 1.69; 4.48]
#> 25  [ 0.38;6.96]
#> 29             3
#> 36            66
#> 33    3.21(1.45)
#> 34    3.28(1.48)
#> 32 [ 2.42; 4.44]
#> 31  [-1.19;6.54]
#> 35             0


# Quantitative statistics with spacing rows (option at.row)

report.quanti(data=data,y="y_numeric",x1="GROUP",
x2="TIMEPOINT",total=TRUE,at.row="TIMEPOINT")

#>    TIMEPOINT Statistics             A             B             C         Total
#> 1         D0          N            30            20            16            66
#> 2         D0  Mean (SD)   -0.93(0.86)   -0.67(1.09)   -1.19(0.92)   -0.92(0.95)
#> 3         D0     Median         -0.82         -0.69         -1.26         -0.86
#> 4         D0    [Q1;Q3] [-1.59;-0.16] [-1.39;-0.06] [-1.62;-0.83] [-1.55;-0.16]
#> 5         D0  [Min;Max]  [-2.34;0.36]  [-2.44;2.10]  [-2.99;0.66]  [-2.99;2.10]
#> 6         D0    Missing             1             1             0             2
#> 7                                                                              
#> 8         D1          N            30            20            16            66
#> 9         D1  Mean (SD)    1.83(1.04)    4.17(1.28)    4.98(0.69)    3.33(1.73)
#> 10        D1     Median          1.78          4.19          5.08          3.57
#> 11        D1    [Q1;Q3] [ 0.94; 2.54] [ 3.23; 4.92] [ 4.58; 5.46] [ 1.78; 4.91]
#> 12        D1  [Min;Max]  [ 0.11;3.88]  [ 1.48;6.19]  [ 3.80;6.23]  [ 0.11;6.23]
#> 13        D1    Missing             1             0             0             1
#> 14                                                                             
#> 15        D2          N            30            20            16            66
#> 16        D2  Mean (SD)    1.97(1.17)    4.04(0.89)    4.90(1.36)    3.32(1.70)
#> 17        D2     Median          1.66          4.19          5.06          3.57
#> 18        D2    [Q1;Q3] [ 1.23; 2.86] [ 3.62; 4.36] [ 4.34; 5.20] [ 1.89; 4.44]
#> 19        D2  [Min;Max]  [-0.18;4.36]  [ 2.03;5.63]  [ 2.39;7.96]  [-0.18;7.96]
#> 20        D2    Missing             1             1             0             2
#> 21                                                                             
#> 22        D3          N            30            20            16            66
#> 23        D3  Mean (SD)    1.78(1.17)    3.81(0.94)    5.07(1.12)    3.15(1.75)
#> 24        D3     Median          1.78          3.63          5.22          3.15
#> 25        D3    [Q1;Q3] [ 0.93; 2.42] [ 3.13; 4.44] [ 4.11; 5.66] [ 1.80; 4.39]
#> 26        D3  [Min;Max]  [-0.16;3.90]  [ 2.46;6.01]  [ 3.16;7.37]  [-0.16;7.37]
#> 27        D3    Missing             0             1             1             2
#> 28                                                                             
#> 29        D4          N            30            20            16            66
#> 30        D4  Mean (SD)    1.83(0.85)    3.80(0.95)    5.17(1.03)    3.22(1.66)
#> 31        D4     Median          1.67          3.83          4.88          3.16
#> 32        D4    [Q1;Q3] [ 1.26; 2.32] [ 3.12; 4.42] [ 4.69; 5.50] [ 1.69; 4.48]
#> 33        D4  [Min;Max]  [ 0.38;3.97]  [ 2.31;5.41]  [ 3.24;6.96]  [ 0.38;6.96]
#> 34        D4    Missing             1             1             1             3
#> 35                                                                             
#> 36        D5          N            30            20            16            66
#> 37        D5  Mean (SD)    2.27(1.20)    3.64(1.19)    4.43(0.98)    3.21(1.45)
#> 38        D5     Median          2.50          3.86          4.57          3.28
#> 39        D5    [Q1;Q3] [ 1.77; 3.21] [ 2.59; 4.60] [ 3.44; 4.97] [ 2.42; 4.44]
#> 40        D5  [Min;Max]  [-1.19;4.31]  [ 0.91;5.12]  [ 2.95;6.54]  [-1.19;6.54]
#> 41        D5    Missing             0             0             0             0


# Add number of subjects in headers (option subjid)

tab=report.quanti(data=data,y="y_numeric",x1="GROUP",
x2="TIMEPOINT",total=TRUE,at.row="TIMEPOINT",subjid="SUBJID")

# Print tab output
tab

#>    TIMEPOINT Statistics      A (N=30)      B (N=21)      C (N=17)  Total (N=68)
#> 1         D0          N            30            20            16            66
#> 2         D0  Mean (SD)   -0.93(0.86)   -0.67(1.09)   -1.19(0.92)   -0.92(0.95)
#> 3         D0     Median         -0.82         -0.69         -1.26         -0.86
#> 4         D0    [Q1;Q3] [-1.59;-0.16] [-1.39;-0.06] [-1.62;-0.83] [-1.55;-0.16]
#> 5         D0  [Min;Max]  [-2.34;0.36]  [-2.44;2.10]  [-2.99;0.66]  [-2.99;2.10]
#> 6         D0    Missing             1             1             0             2
#> 7                                                                              
#> 8         D1          N            30            20            16            66
#> 9         D1  Mean (SD)    1.83(1.04)    4.17(1.28)    4.98(0.69)    3.33(1.73)
#> 10        D1     Median          1.78          4.19          5.08          3.57
#> 11        D1    [Q1;Q3] [ 0.94; 2.54] [ 3.23; 4.92] [ 4.58; 5.46] [ 1.78; 4.91]
#> 12        D1  [Min;Max]  [ 0.11;3.88]  [ 1.48;6.19]  [ 3.80;6.23]  [ 0.11;6.23]
#> 13        D1    Missing             1             0             0             1
#> 14                                                                             
#> 15        D2          N            30            20            16            66
#> 16        D2  Mean (SD)    1.97(1.17)    4.04(0.89)    4.90(1.36)    3.32(1.70)
#> 17        D2     Median          1.66          4.19          5.06          3.57
#> 18        D2    [Q1;Q3] [ 1.23; 2.86] [ 3.62; 4.36] [ 4.34; 5.20] [ 1.89; 4.44]
#> 19        D2  [Min;Max]  [-0.18;4.36]  [ 2.03;5.63]  [ 2.39;7.96]  [-0.18;7.96]
#> 20        D2    Missing             1             1             0             2
#> 21                                                                             
#> 22        D3          N            30            20            16            66
#> 23        D3  Mean (SD)    1.78(1.17)    3.81(0.94)    5.07(1.12)    3.15(1.75)
#> 24        D3     Median          1.78          3.63          5.22          3.15
#> 25        D3    [Q1;Q3] [ 0.93; 2.42] [ 3.13; 4.44] [ 4.11; 5.66] [ 1.80; 4.39]
#> 26        D3  [Min;Max]  [-0.16;3.90]  [ 2.46;6.01]  [ 3.16;7.37]  [-0.16;7.37]
#> 27        D3    Missing             0             1             1             2
#> 28                                                                             
#> 29        D4          N            30            20            16            66
#> 30        D4  Mean (SD)    1.83(0.85)    3.80(0.95)    5.17(1.03)    3.22(1.66)
#> 31        D4     Median          1.67          3.83          4.88          3.16
#> 32        D4    [Q1;Q3] [ 1.26; 2.32] [ 3.12; 4.42] [ 4.69; 5.50] [ 1.69; 4.48]
#> 33        D4  [Min;Max]  [ 0.38;3.97]  [ 2.31;5.41]  [ 3.24;6.96]  [ 0.38;6.96]
#> 34        D4    Missing             1             1             1             3
#> 35                                                                             
#> 36        D5          N            30            20            16            66
#> 37        D5  Mean (SD)    2.27(1.20)    3.64(1.19)    4.43(0.98)    3.21(1.45)
#> 38        D5     Median          2.50          3.86          4.57          3.28
#> 39        D5    [Q1;Q3] [ 1.77; 3.21] [ 2.59; 4.60] [ 3.44; 4.97] [ 2.42; 4.44]
#> 40        D5  [Min;Max]  [-1.19;4.31]  [ 0.91;5.12]  [ 2.95;6.54]  [-1.19;6.54]
#> 41        D5    Missing             0             0             0             0


#Getting raw output
tab$raw.output

#>    GROUP TIMEPOINT  N       mean        sd     median         q1          q3
#> 1      A        D0 30 -0.9307357 0.8565778 -0.8240857 -1.5897260 -0.15709410
#> 2      A        D1 30  1.8306336 1.0382655  1.7849384  0.9437696  2.54138105
#> 3      A        D2 30  1.9661080 1.1726220  1.6636286  1.2328514  2.85721616
#> 4      A        D3 30  1.7770473 1.1651704  1.7760930  0.9280940  2.42451639
#> 5      A        D4 30  1.8300697 0.8458283  1.6695989  1.2586428  2.31952759
#> 6      A        D5 30  2.2737568 1.2022079  2.5031150  1.7672201  3.21282837
#> 7      B        D0 20 -0.6717750 1.0880504 -0.6920881 -1.3877254 -0.05550822
#> 8      B        D1 20  4.1672970 1.2845806  4.1915179  3.2302158  4.91901508
#> 9      B        D2 20  4.0416510 0.8870067  4.1866387  3.6223517  4.36392654
#> 10     B        D3 20  3.8079135 0.9363071  3.6277336  3.1310529  4.43941943
#> 11     B        D4 20  3.8006831 0.9476812  3.8276619  3.1157893  4.42371324
#> 12     B        D5 20  3.6444439 1.1914445  3.8610892  2.5854548  4.60284254
#> 13     C        D0 16 -1.1877651 0.9236343 -1.2579996 -1.6213937 -0.83193494
#> 14     C        D1 16  4.9845689 0.6880475  5.0835021  4.5783264  5.45922166
#> 15     C        D2 16  4.8983106 1.3551805  5.0572353  4.3443597  5.20411764
#> 16     C        D3 16  5.0747127 1.1248254  5.2219356  4.1143592  5.65976546
#> 17     C        D4 16  5.1732328 1.0307377  4.8832664  4.6868444  5.50058498
#> 18     C        D5 16  4.4305195 0.9775492  4.5737580  3.4391191  4.96627414
#>           min       max missing
#> 1  -2.3378197 0.3627234       1
#> 2   0.1088366 3.8830349       1
#> 3  -0.1827715 4.3610839       1
#> 4  -0.1602072 3.9024250       0
#> 5   0.3848371 3.9690577       1
#> 6  -1.1927967 4.3082425       0
#> 7  -2.4440777 2.1022772       1
#> 8   1.4755850 6.1949969       0
#> 9   2.0275024 5.6293059       1
#> 10  2.4609821 6.0147632       1
#> 11  2.3057726 5.4060729       1
#> 12  0.9063581 5.1240373       0
#> 13 -2.9922843 0.6631568       0
#> 14  3.8018054 6.2253186       0
#> 15  2.3855790 7.9626442       0
#> 16  3.1627164 7.3666613       1
#> 17  3.2383776 6.9611225       1
#> 18  2.9491903 6.5438513       0