Add space to the results of a statistic table

spacetable Add space at some interval in a table of statistics

spacetable(tab, at.row = 6)

Arguments

tab
a data.frame. Normally, the output part of a desc object

at.row
Numeric (between 1 and ncol(tab)) or Character (a column name of tab)

Value

A data.frame with some space between levels of the variable specified by at.row

Details

Used internally into report.lsmeans, report.quanti and report.quali

See also

Examples


data(data)
spacetable(report.quali(data=data,y="y_logistic",x1="GROUP",
x2="TIMEPOINT")$output,at.row="TIMEPOINT")

#>    TIMEPOINT Levels    Statistics          A          B          C
#> 1         D0      0  n (column %) 11(36.67%) 11(55.00%)  7(43.75%)
#> 2         D0      1  n (column %) 18(60.00%)  8(40.00%)  7(43.75%)
#> 3         D0         Missing n(%)   1(3.33%)   1(5.00%)  2(12.50%)
#> 4                                                                 
#> 5                                                                 
#> 6                                                                 
#> 7                                                                 
#> 8                                                                 
#> 9                                                                 
#> 10                                                                
#> 11        D1      0  n (column %)  7(23.33%) 13(65.00%)  8(50.00%)
#> 12        D1      1  n (column %) 21(70.00%)  7(35.00%)  7(43.75%)
#> 13        D1         Missing n(%)   2(6.67%)      0(0%)   1(6.25%)
#> 14                                                                
#> 15        D2      0  n (column %) 18(60.00%)  7(35.00%) 11(68.75%)
#> 16        D2      1  n (column %) 12(40.00%) 13(65.00%)  5(31.25%)
#> 17        D2         Missing n(%)      0(0%)      0(0%)      0(0%)
#> 18                                                                
#> 19        D3      0  n (column %) 11(36.67%) 10(50.00%)  7(43.75%)
#> 20        D3      1  n (column %) 19(63.33%) 10(50.00%)  9(56.25%)
#> 21        D3         Missing n(%)      0(0%)      0(0%)      0(0%)
#> 22                                                                
#> 23        D4      0  n (column %) 18(60.00%) 12(60.00%)  6(37.50%)
#> 24        D4      1  n (column %) 12(40.00%)  8(40.00%)  8(50.00%)
#> 25        D4         Missing n(%)      0(0%)      0(0%)  2(12.50%)
#> 26                                                                
#> 27        D5      0  n (column %) 14(46.67%)  7(35.00%)  8(50.00%)
#> 28        D5      1  n (column %) 15(50.00%) 13(65.00%)  8(50.00%)
#> 29        D5         Missing n(%)   1(3.33%)      0(0%)      0(0%)